junction-core: switch to delta
