+++ crates/junction-core/src/xds/cache.rs
@@ -1,162 +1,127 @@
-// This module is a cache that handles SotW XDS behavior. It's the guts of an
-// ADS connection and tracks the current state of XDS resoureces, the references
-// between resources, and builds internal client config on the fly. If you need
-// to add or modify any XDS behavior at all, it's like you'll end up here.
-//
-// This cache is built entirely around being single-writer, even thought it's
-// multi-reader and safe for concurrent reads. If you'd like to change that,
-// it's likely you're going to rebuild the internals of the cache entirely.
-//
-// # Reference Tracking is XDS Subscription Tracking
-//
-// XDS SotW only ever issues deletes for LDS and CDS resources, so one of the
-// most important jobs a cache has is tracking references to delete RDS/EDS
-// objects when they're no longer referenced. Tracking the list of referenced
-// names is also exactly what we need to be doing for subscription tracking -
-// any client using this cache should be subscribing to all RDS resources
-// referenced by existing LDS resources, all CDS resources referenced by LDS and
-// RDS resources, and so on.
-//
-// The current implementation manages these references as a `petgraph` graph
-// owned by the cache's single writer, and not shared with any of the readers.
-// This means that reference data can be stored relatively cheaply (without
-// duplicating XDS protobufs) and modified without any coordination between
-// reader and writer threads/tasks. It does however mean that there are two
-// places to track state, and the writer is now responsible for keeping them
-// in sync.
-//
-// # Reference Tracking is Garbage Collection
-//
-// Tracking a graph of objects that reference each other and removing the
-// unused ones should sound a lot like garbage collection to you - it is
-// exactly garbage collection.
-//
-// Using a graph internally to track object references means that it's
-// relatively easy to run a simple mark-and-sweep over the current state of the
-// graph, and having a single writer own the reference graph means that writes
-// pay the cost of collecting XDS garbage while readers can keep reading
-// uninterrupted. In practice, we expect the number of XDS objects to be
-// relatively small, and this cost to be low, but this is the right tradeoff
-// even if collection does become more expensive.
-//
-// # User Input is GC Roots
-//
-// This cache models our entire interaction with ADS, so we need a notion of
-// what to request on behalf of all the clients reading config from it. As of
-// now, that generally takes the form of LDS resources - a client makes a
-// request to a URL, and the hostname of that URL is now a Listener we'd like to
-// subscribe to.
-//
-// That naturally makes LDS/CDS resources our GC roots - when someone explicitly
-// subscribes to a Lister (and maybe sets up default routes with targets)
-// Listener names become roots in our GC graph. We follow all references
-// downstream to other XDS objects to decide what to drop and keep, and even if
-// the ADS server tells us those names are temporarily gone, the cache should
-// keep trying to subscribe to them - after all, a user has expressed interest.
-//
-// All of this assumes that there are no wildcard subscriptions - the huge
-// assumption here is that subscriptions are going to come in as explicit DNS
-// names, and not as wildcards - it doesn't really mean anything to try to make
-// an http request to `*.foo.local`. If this changes, we'll have to re-evaluate
-// our model of the world.
-//
-// # TODO
-//
-// - Track incoming resource versions and last update. When dumping resources
-//   for CSDS, there's the opportunity to show both of those pieces of info to
-//   a caller, which should be extremely useful for debugging.
-//
-// - Use the resource graph to track when resources were requested and whether
-//   or not they should be considered missing. The XDS protocol documentation
-//   recommends a 15 second timeout. This probably involves also inserting
-//   markers/tombstones in the data cache.
-//   https://www.envoyproxy.io/docs/envoy/latest/api-docs/xds_protocol#knowing-when-a-requested-resource-does-not-exist.
-//
+//! This package contains the state of the Junction xDS client that survives
+//! between individual ADS connections. The cache is built with incremental xDS
+//! in mind, but could potentially be used for a State-of-the-World connection
+//! or to generate individual Fetch requests if there's a reason to not use the
+//! streaming variants of the protocol.
+//!
+//! The cache in this package is built around having a single writer and any
+//! number of concurrent readers (see [Cache::reader]), with the idea that
+//! updates are rare relative to reads and that updates should never block
+//! reads.
+//!
+//! The writer handles resource updates, interperets xDS as Junction types, and
+//! manages the relationships between resources. As resources are handled and
+//! inserted into the cache, they become immediately available to readers.
+//!
+//! # Junction Resources
+//!
+//! This package contains a set of Junction-flavored-xDS resources, that include
+//! raw xDS and Junction-specific information about a resource, which is a
+//! superset of the `junction-api` types a Resource represents. For example, an
+//! xDS `RouteConfig` in this package contains both the `Route` that the xDS
+//! represents along with the pre-computed xDS clusters that the Route
+//! references.
+//!
+//! # Reading from a Cache
+//!
+//! All readers are expected to use a [CacheReader] to read from a cache.
+//! Readers are cheap to clone, and should be freely handed out to each
+//! individual task that wants to read from a cache.
+//!
+//! # Writing to a Cache
+//!
+//! The owner of a Cache is expected to call [Cache::insert], [Cache::subscribe]
+//! and [Cache::unsubscribe] in batches, and then periodically call
+//! [Cache::collect] to remove unused resources and generate a set of resources
+//! to add/remove via xDS.
+//!
+//! The current state of the cache can also be inspected by with method like
+//! [Cache::versions] or [Cache::initial_subscriptions]. These methods should
+//! generally be called after a call to [Cache::collect] - if called between a
+//! change to cache state and a collection, they may return references to
+//! resources that are no longer connected to the rest of the resource graph.
+//!
+//! ## Subscriptions and Garbage Collection
+//!
+//! One of the primary jobs of the cache writer is to track relationships
+//! between resources and determine what resources should be requested on any
+//! individual connection. This is a little more complex than just tracking what
+//! resources have been explicitly subscribed to with a call to
+//! [Cache::subscribe]. A `Listener` subscription creates a subscription to a
+//! `RouteConfiguration`, which generates subscriptions to some number of
+//! `Cluster`s, and so on. Unsubscribing is also not straight forward for the
+//! same reasons - while the caller may not explicitly care about a `Cluster`
+//! any more, it may be a dependency of an existing `RouteConfiguration`, and
+//! shouldn't be removed from cache even though there is no longer explicit
+//! interest in it.
+//!
+//! To handle resources properly, a Cache builds a reference graph between xDS
+//! Resources and treats this like a garbage collection problem. Any resource
+//! that's been explicitly subcribed to is treated as a root of the reference
+//! graph. If the resource type accepts wildcard subscriptions, any resource
+//! inserted as a wildcard is also treated as a root.
+//!
+//! Using the root set, garbage collection is done with a very simple
+//! mark-and-sweep approach. [Cache::collect] forces a collection, and returns
+//! the set of changes to resource subscriptions that happened during the last
+//! batch of inserts and during garbage collection.
+//!
+//! ## Subscription Changes
+//!
+//! As resources get added and removed from a [Cache], it builds up a list of
+//! changes to the subscription graph. When [Cache::collect] is called it
+//! returns the set of added and removed subscriptions for each resource type in
+//! the graph, which includes the set of resources that may have been removed
+//! during garbage collection.
+//!
+//! Tracking changes means it's easy to generate ADS updates for those resource
+//! types - changes for a resource type new requests need to be generated to
+//! send updates to the server about what the connection is now interested in.
+//!
+//! Cluster changes can also create a dependency on client DNS. Changes to DNS
+//! hostnames of interest are also tracked and returned.
+
+use std::{
+    collections::{BTreeSet, HashMap},
+    str::FromStr,
+    sync::Arc,
+};
 
 use crossbeam_skiplist::SkipMap;
 use enum_map::EnumMap;
-use junction_api::http::Route;
-use junction_api::{backend::BackendId, Hostname, Service};
+use junction_api::{backend::BackendId, http::Route, Hostname};
 use petgraph::{
     graph::{DiGraph, NodeIndex},
-    visit::{self, Visitable},
+    visit::{EdgeRef, Visitable},
     Direction,
 };
-use prost::Name;
-use std::collections::BTreeSet;
-use std::str::FromStr;
-use std::sync::Arc;
 use tokio::sync::Notify;
 use xds_api::pb::envoy::config::{
-    cluster::v3::{self as xds_cluster},
-    endpoint::v3::{self as xds_endpoint},
-    listener::v3::{self as xds_listener},
-    route::v3::{self as xds_route},
+    cluster::v3 as xds_cluster, endpoint::v3 as xds_endpoint, listener::v3 as xds_listener,
+    route::v3 as xds_route,
 };
 use xds_api::pb::google::protobuf;
 
-// collect garbage like a little tracing gc.
-//
-// this traversal also asserts that the GC graph is a DAG, and will
-// panic if it finds cycles. the ref graph is directed, and there
-// should be no cycles by design - there are no self-type references,
-// and no references "up" the xds type hierarchy.
-//
-// all listeners are GC roots. walk the graph once to find them here,
-// instead of storing them. storing them involves keeping a secondary
-// index from name to graph indices, but indices are unstable.
-//
-// this is mostly a handful of DFS passes on the graph, but with an
-// early exit if we've already marked a node.
+use crate::{endpoints::EndpointGroup, BackendLb};
 
-use crate::{BackendLb, ConfigCache, EndpointGroup};
-
-use super::resources::{
-    ApiListener, ApiListenerData, Cluster, LoadAssignment, ResourceError, ResourceName,
-    ResourceType, ResourceTypeSet, ResourceVec, RouteConfig, RouteConfigData,
+use super::{
+    resources::{
+        ApiListener, ApiListenerData, Cluster, LoadAssignment, ResourceError, RouteConfig,
+        RouteConfigData,
+    },
+    ConfigCache, DnsUpdates, ResourceType, ResourceVec, ResourceVersion, XdsConfig,
 };
-use super::{ResourceVersion, XdsConfig};
 
-#[derive(Debug, Clone)]
-struct CacheEntry<T> {
-    version: ResourceVersion,
-    last_error: Option<(ResourceVersion, ResourceError)>,
-    data: Option<T>,
-}
-
-impl<T: CacheEntryData> CacheEntry<T> {}
-
-trait CacheEntryData {
-    type Xds;
-
-    fn xds(&self) -> &Self::Xds;
-}
-
-macro_rules! impl_cache_entry {
-    ($entry_ty:ty, $xds_ty:ty) => {
-        impl CacheEntryData for $entry_ty {
-            type Xds = $xds_ty;
-
-            fn xds(&self) -> &$xds_ty {
-                &self.xds
-            }
-        }
-    };
-}
-
-impl_cache_entry!(ApiListener, xds_listener::Listener);
-impl_cache_entry!(RouteConfig, xds_route::RouteConfiguration);
-impl_cache_entry!(Cluster, xds_cluster::Cluster);
-impl_cache_entry!(LoadAssignment, xds_endpoint::ClusterLoadAssignment);
-
+/// A concurrent map of resources, bundled together with an Arc<Notify> so that
+/// callers can wait on changes.
 #[derive(Debug)]
 struct ResourceMap<T> {
     changed: Arc<Notify>,
     map: SkipMap<String, CacheEntry<T>>,
 }
 
-// NOTE: manually derived because the Derive macro requires T: Default, and
-// this impl doesn't care - an empty map doesn't need a T.
+// NOTE: manually derived because the Derive macro requires `T: Default`, and
+// this impl shouldn't care - an empty map doesn't need a T.
 impl<T> Default for ResourceMap<T> {
     fn default() -> Self {
         Self {
@@ -221,13 +186,35 @@ impl<T: Send + 'static> ResourceMap<T> {
         self.map.iter().map(ResourceEntry)
     }
 
-    fn names(&self) -> impl Iterator<Item = String> + '_ {
-        self.map.iter().map(|e| e.key().clone())
+    fn has_data(&self, k: &str) -> bool {
+        match self.get(k) {
+            None => false,
+            Some(entry) => entry.data().is_some(),
+        }
     }
 
-    fn remove(&self, name: &str) {
-        self.map.remove(name);
+    fn versions(&self) -> HashMap<String, String> {
+        let mut versions = HashMap::new();
+        for entry in self.iter() {
+            if entry.data().is_none() {
+                continue;
+            };
+            let Some(version) = entry.version() else {
+                continue;
+            };
+
+            let name = entry.name().to_string();
+            let version = version.to_string();
+            versions.insert(name, version);
+        }
+
+        versions
+    }
+
+    fn remove(&self, name: &str) -> Option<ResourceEntry<T>> {
+        let entry = self.map.remove(name);
         self.changed.notify_waiters();
+        entry.map(ResourceEntry)
     }
 
     fn remove_all<I>(&self, names: I)
@@ -235,8 +222,7 @@ impl<T: Send + 'static> ResourceMap<T> {
         I: IntoIterator<Item: AsRef<str>>,
     {
         for name in names {
-            self.map.remove(name.as_ref());
-            self.changed.notify_waiters();
+            self.remove(name.as_ref());
         }
     }
 }
@@ -250,9 +236,21 @@ where
         self.map.insert(
             name,
             CacheEntry {
-                version,
+                version: Some(version),
                 last_error: None,
                 data: Some(t),
+            },
+        );
+        self.changed.notify_waiters();
+    }
+
+    fn insert_tombstone(&self, name: String) {
+        self.map.insert(
+            name,
+            CacheEntry {
+                version: None,
+                last_error: None,
+                data: None,
             },
         );
         self.changed.notify_waiters();
@@ -275,7 +273,7 @@ where
                 self.map.insert(
                     name,
                     CacheEntry {
-                        version: ResourceVersion::default(),
+                        version: None,
                         last_error: Some((version, error.into())),
                         data: None,
                     },
@@ -285,9 +283,7 @@ where
         }
     }
 
-    // TDODO: should this also compare version? for Cluster and Listener it'd
-    // mean a decent amount of churn replacing an identical resource on every
-    // update.
+    // TDODO: should this compare version to short-circuit full eq?
     fn is_changed(&self, name: &str, t: &X) -> bool {
         let Some(entry) = self.map.get(name) else {
             return true;
@@ -300,6 +296,8 @@ where
     }
 }
 
+/// Wrap a crossbeam_skiplist Entry to make some signatures and reference things
+/// less gnarly.
 struct ResourceEntry<'a, T>(crossbeam_skiplist::map::Entry<'a, String, CacheEntry<T>>);
 
 impl<T> ResourceEntry<'_, T> {
@@ -307,8 +305,8 @@ impl<T> ResourceEntry<'_, T> {
         self.0.key()
     }
 
-    fn version(&self) -> &ResourceVersion {
-        &self.0.value().version
+    fn version(&self) -> Option<&ResourceVersion> {
+        self.0.value().version.as_ref()
     }
 
     fn last_error(&self) -> Option<&(ResourceVersion, ResourceError)> {
@@ -320,14 +318,797 @@ impl<T> ResourceEntry<'_, T> {
     }
 }
 
-/// A read-only handle to a [Cache]. `CacheReader`s are meant to passed around
-/// and shared and are cheap to clone.
+#[derive(Clone, Debug)]
+struct CacheEntry<T> {
+    version: Option<ResourceVersion>,
+    last_error: Option<(ResourceVersion, ResourceError)>,
+    data: Option<T>,
+}
+
+impl<T> Default for CacheEntry<T> {
+    fn default() -> Self {
+        Self {
+            version: None,
+            last_error: None,
+            data: None,
+        }
+    }
+}
+
+impl<T: CacheEntryData> CacheEntry<T> {}
+
+// TODO: kill this, move xds into CacheEntry
+trait CacheEntryData {
+    type Xds;
+
+    fn xds(&self) -> &Self::Xds;
+}
+
+macro_rules! impl_cache_entry {
+    ($entry_ty:ty, $xds_ty:ty) => {
+        impl CacheEntryData for $entry_ty {
+            type Xds = $xds_ty;
+
+            fn xds(&self) -> &$xds_ty {
+                &self.xds
+            }
+        }
+    };
+}
+
+impl_cache_entry!(ApiListener, xds_listener::Listener);
+impl_cache_entry!(RouteConfig, xds_route::RouteConfiguration);
+impl_cache_entry!(Cluster, xds_cluster::Cluster);
+impl_cache_entry!(LoadAssignment, xds_endpoint::ClusterLoadAssignment);
+
+#[derive(Debug, Default)]
+struct CacheData {
+    listeners: ResourceMap<ApiListener>,
+    route_configs: ResourceMap<RouteConfig>,
+    clusters: ResourceMap<Cluster>,
+    load_assignments: ResourceMap<LoadAssignment>,
+}
+
+/// The set of subscription changes for a single resource type.
+#[derive(Debug, Default, PartialEq, Eq)]
+pub(crate) struct Changes {
+    /// The names of any newly added subscriptions. Newly added subscriptions
+    /// may or may not be in cache, and must be requested from the xDS server.
+    pub(crate) added: BTreeSet<String>,
+
+    /// The names of any removed subscriptions. Callers should assume that any
+    /// removed names are also removed from cache.
+    pub(crate) removed: BTreeSet<String>,
+}
+
+impl Changes {
+    pub(crate) fn is_empty(&self) -> bool {
+        self.added.is_empty() && self.removed.is_empty()
+    }
+}
+
+/// Subscription tracking for xDS. This struct exists with the expectation that
+/// it's owned by a single [Cache] and is being used to track the relationships
+/// between resources in that cache. No introspection of xDS is done here.
+///
+/// # Wildcard Mode
+#[derive(Debug)]
+struct Subscriptions {
+    subs: DiGraph<SubscriptionInfo, ()>,
+    changes: EnumMap<ResourceType, Changes>,
+    wildcard: EnumMap<ResourceType, bool>,
+}
+
+impl Default for Subscriptions {
+    fn default() -> Self {
+        let mut wildcard = EnumMap::default();
+        for rtype in ResourceType::all() {
+            wildcard[*rtype] = rtype.supports_wildcard()
+        }
+
+        Self {
+            subs: Default::default(),
+            changes: Default::default(),
+            wildcard,
+        }
+    }
+}
+
+// TODO: add insertion times and TTLs
+#[derive(Debug)]
+struct SubscriptionInfo {
+    // the type of the resource
+    resource_type: ResourceType,
+
+    // the name of the resource
+    name: String,
+
+    // true if there is explicit interest in this resource via subscribe.
+    explicit: bool,
+
+    // true if this subscription was added by inserting a resource that didn't
+    // have an existing subscription.
+    wildcard: bool,
+}
+
+impl Subscriptions {
+    fn explicit(&self, rtype: ResourceType) -> impl Iterator<Item = &str> + '_ {
+        self.subs
+            .node_weights()
+            .filter(move |w| w.resource_type == rtype && !w.wildcard)
+            .map(|w| &w.name[..])
+    }
+
+    fn subscribe(&mut self, rtype: ResourceType, name: &str) {
+        // explicit subscription means never a wildcard
+        let sub = self.find_or_create(rtype, name, false);
+        self.subs[sub].explicit = true;
+    }
+
+    fn unsubscribe(&mut self, rtype: ResourceType, name: &str) {
+        if let Some(sub) = self.find(rtype, name) {
+            self.subs[sub].explicit = false;
+        }
+    }
+
+    fn remove(&mut self, rtype: ResourceType, name: &str) {
+        if let Some(sub) = self.find_subcribed(rtype, name) {
+            self.reset_refs(sub);
+        }
+    }
+
+    #[inline]
+    fn clear_changes(&mut self, rtype: ResourceType, name: &str) {
+        self.changes[rtype].added.remove(name);
+        self.changes[rtype].removed.remove(name);
+    }
+
+    /// safety: must be called with a valid NodeIndex
+    fn remove_sub(&mut self, sub: NodeIndex) {
+        let sub = self.subs.remove_node(sub).unwrap();
+        self.changes[sub.resource_type].removed.insert(sub.name);
+    }
+
+    fn find_subcribed(&mut self, rtype: ResourceType, name: &str) -> Option<NodeIndex> {
+        // if this is a wildcard subscription, any name is subscribed. create it
+        // and move on.
+        if self.wildcard[rtype] {
+            // if this is a new node and is a wildcard, don't track the
+            // side-effect of adding the subscription. we've gotten the resource
+            // already.
+            let sub = self.find_or_create(rtype, name, true);
+            return Some(sub);
+        }
+
+        self.find(rtype, name)
+    }
+
+    fn find(&self, rtype: ResourceType, name: &str) -> Option<NodeIndex> {
+        self.subs.node_indices().find(|idx| {
+            let sub = &self.subs[*idx];
+            sub.resource_type == rtype && sub.name == name
+        })
+    }
+
+    fn find_or_create(&mut self, rtype: ResourceType, name: &str, wildcard: bool) -> NodeIndex {
+        match self.find(rtype, name) {
+            Some(idx) => idx,
+            None => {
+                let idx = self.subs.add_node(SubscriptionInfo {
+                    name: name.to_string(),
+                    resource_type: rtype,
+                    explicit: false,
+                    wildcard,
+                });
+
+                // track that this is a newly created sub
+                if !wildcard {
+                    self.changes[rtype].added.insert(name.to_string());
+                }
+
+                idx
+            }
+        }
+    }
+
+    /// Remove all of of a subscription node's outgoing edges. Any references to
+    /// this node are left untouched.
+    ///
+    /// Safety: can be called with NodeIndexes outstanding, should not modify
+    /// NodeIndexes and make them unstable.
+    fn reset_refs(&mut self, sub: NodeIndex) {
+        let out_refs: Vec<_> = self
+            .subs
+            .edges_directed(sub, Direction::Outgoing)
+            .map(|edge_ref| edge_ref.id())
+            .collect();
+
+        for out_ref in out_refs {
+            self.subs.remove_edge(out_ref);
+        }
+    }
+
+    /// Add a reference from `from_sub` to a node of with a given resource type
+    /// and name. Creates the destination subscription if it doesn't already
+    /// exist.
+    fn add_ref(&mut self, from_sub: NodeIndex, rtype: ResourceType, name: &str) {
+        // when adding a reference that creates a new subscription, even if the
+        // destination type is a wildcard, we want a non-wildcard reference to
+        // it so that the cache can switch to explicit mode and keep this
+        // reference.
+        let to_sub = self.find_or_create(rtype, name, false);
+        self.subs.add_edge(from_sub, to_sub, ());
+    }
+
+    fn collect(&mut self) {
+        use petgraph::visit::{Control, DfsEvent};
+
+        // walk the GC graph, keeping the set of the reachable nodes.
+        //
+        // lean on petgraph's Control to only visit each node once - because
+        // the ref graph must be a DAG, we can skip marking nodes twice and
+        // emit Control::Prune every time we see a node we've already seen.
+        let mut reachable = self.subs.visit_map();
+        petgraph::visit::depth_first_search(&self.subs, self.gc_roots(), |event| -> Control<()> {
+            if let DfsEvent::Discover(n, _) = event {
+                if reachable.contains(n.index()) {
+                    return Control::Prune;
+                }
+                reachable.insert(n.index());
+            };
+
+            Control::Continue
+        });
+
+        // remove all unreachalbe nodes from the graph
+        //
+        // safety: remove_node invalidates the last index in the graph when
+        // called. walking the indices backwards means that we're guaranteed
+        // to not be invalidating an index in the reachable set that we haven't
+        // touched yet.
+        for idx in self.subs.node_indices().rev() {
+            if reachable.contains(idx.index()) {
+                continue;
+            }
+
+            self.remove_sub(idx);
+        }
+    }
+
+    fn is_gc_root(&self, node: NodeIndex) -> bool {
+        let sub_data = &self.subs[node];
+        sub_data.explicit || sub_data.wildcard
+    }
+
+    fn gc_roots(&self) -> Vec<NodeIndex> {
+        self.subs
+            .node_indices()
+            .filter(|idx| self.is_gc_root(*idx))
+            .collect()
+    }
+}
+
+/// A persistent cache of Junction xDS state. See the module documentation for
+/// more information on what's in a cache and how to use one.
+#[derive(Debug, Default)]
+pub(super) struct Cache {
+    subs: Subscriptions,
+    data: Arc<CacheData>,
+    dns: DnsUpdates,
+}
+
+impl Cache {
+    /// Create a new read-only handle to this cache.
+    ///
+    /// Read handles are cheap, and intended to be created and shared across
+    /// multiple threads and tasks.
+    pub(super) fn reader(&self) -> CacheReader {
+        CacheReader {
+            data: self.data.clone(),
+        }
+    }
+
+    /// Set wildcard mode for a resource type. Will have no effect if the
+    /// resource doesn't support wildcard mode.
+    #[cfg(test)]
+    pub(crate) fn set_wildcard(&mut self, rtype: ResourceType, wildcard: bool) {
+        if !rtype.supports_wildcard() {
+            return;
+        }
+        self.subs.wildcard[rtype] = wildcard;
+    }
+
+    /// Check whether a resource type is in wildcard mode.
+    pub(crate) fn is_wildcard(&self, rtype: ResourceType) -> bool {
+        self.subs.wildcard[rtype]
+    }
+
+    /// Subscribe to a resource by name.
+    pub(crate) fn subscribe(&mut self, rtype: ResourceType, name: &str) {
+        self.subs.subscribe(rtype, name);
+    }
+
+    /// Unsubscribe from a resource by name.
+    pub(crate) fn unsubscribe(&mut self, rtype: ResourceType, name: &str) {
+        self.subs.unsubscribe(rtype, name);
+    }
+
+    /// Subscribe to a DNS name.
+    pub(crate) fn subscribe_dns(&mut self, hostname: Hostname, port: u16) {
+        self.dns.add.insert((hostname, port));
+    }
+
+    /// Unsubscribe from a DNS name.
+    pub(crate) fn unsubscribe_dns(&mut self, hostname: Hostname, port: u16) {
+        self.dns.remove.insert((hostname, port));
+    }
+
+    /// Return the current list of subscriptions for this resource type.
+    #[cfg(test)]
+    pub(crate) fn subscriptions(&self, rtype: ResourceType) -> Vec<String> {
+        self.subs.explicit(rtype).map(|s| s.to_string()).collect()
+    }
+
+    pub(crate) fn dns_names(&self) -> impl Iterator<Item = (Hostname, u16)> + '_ {
+        self.data
+            .clusters
+            .iter()
+            .filter_map(|e| e.data().and_then(|c| c.dns_name()))
+    }
+
+    /// Return the list of resources the cache has a registered subscription for
+    /// but contains no data for.
+    pub(crate) fn initial_subscriptions(&self, rtype: ResourceType) -> Vec<String> {
+        macro_rules! missing_from {
+            ($m:expr) => {
+                self.subs
+                    .explicit(rtype)
+                    .filter(|k| !$m.has_data(k))
+                    .map(|s| s.to_string())
+                    .collect()
+            };
+        }
+
+        match rtype {
+            ResourceType::Cluster => missing_from!(self.data.clusters),
+            ResourceType::ClusterLoadAssignment => missing_from!(self.data.load_assignments),
+            ResourceType::Listener => missing_from!(self.data.listeners),
+            ResourceType::RouteConfiguration => missing_from!(self.data.route_configs),
+        }
+    }
+
+    /// Get the versions of all resources currently in cache.
+    pub(crate) fn versions(&self, rtype: ResourceType) -> HashMap<String, String> {
+        match rtype {
+            ResourceType::Cluster => self.data.clusters.versions(),
+            ResourceType::ClusterLoadAssignment => self.data.load_assignments.versions(),
+            ResourceType::Listener => self.data.listeners.versions(),
+            ResourceType::RouteConfiguration => self.data.route_configs.versions(),
+        }
+    }
+
+    /// Garbage collect the cache.
+    ///
+    /// Returns the set of subscription changes and dns updates that have
+    /// happened since the last call to `collect`.
+    ///
+    /// See the module docs for more on how to use a cache and when its
+    /// appropriate to call `collect`.
+    pub(crate) fn collect(&mut self) -> (EnumMap<ResourceType, Changes>, DnsUpdates) {
+        // first, garbage collect and accumulate all of the pending changes.
+        self.subs.collect();
+        let changes = std::mem::take(&mut self.subs.changes);
+
+        // remove actual resource data in reverse make-before-break order
+        // based on the change set.
+        macro_rules! remove_all {
+            ($field:ident, $rtype:expr) => {
+                self.data.$field.remove_all(&changes[$rtype].removed)
+            };
+        }
+        remove_all!(route_configs, ResourceType::RouteConfiguration);
+        remove_all!(listeners, ResourceType::Listener);
+        remove_all!(load_assignments, ResourceType::ClusterLoadAssignment);
+
+        // when removing clusters based on changes, we also have to remove
+        // any DNS names for removed clusters.
+        let mut dns = std::mem::take(&mut self.dns);
+        for cluster_name in &changes[ResourceType::Cluster].removed {
+            if let Some(entry) = self.data.clusters.remove(cluster_name) {
+                let dns_name = entry.data().and_then(|c| c.dns_name());
+                if let Some(dns_name) = dns_name {
+                    dns.remove.insert(dns_name);
+                }
+            }
+        }
+
+        (changes, dns)
+    }
+
+    /// Insert new resources into cache.
+    ///
+    /// Returns an error for each resource that could not be inserted.
+    /// Successfully parsed resources are visible to readers as soon as they're
+    /// inserted.
+    pub(crate) fn insert(&mut self, resources: ResourceVec) -> Vec<ResourceError> {
+        match resources {
+            ResourceVec::Cluster(clusters) => self.insert_clusters(clusters),
+            ResourceVec::ClusterLoadAssignment(clas) => self.insert_load_assignments(clas),
+            ResourceVec::Listener(listeners) => self.insert_listeners(listeners),
+            ResourceVec::RouteConfiguration(rcs) => self.insert_route_configs(rcs),
+        }
+    }
+
+    /// Remove a list of resources from cache by name.
+    ///
+    /// Removing a resource immediately removes data from the cache, but doesn't
+    /// change the cache's subscription interest - a caller has told us the resource
+    /// data no longer exists, not that the cache shouldn't care about it anymore.
+    ///
+    /// Removing a resource does modify the subscription graph - when removing a
+    /// resource, we need to invalidate any references to other resource types
+    /// that it may have subscribed us to. Note that newly-orphaned resources
+    /// may not be fully removed from cache until the next call to
+    /// [Cache::collect].
+    pub(crate) fn remove(&mut self, rtype: ResourceType, names: &[String]) {
+        macro_rules! tombstone_all {
+            ($data:ident, $rtype:expr, $names:expr) => {{
+                for name in $names {
+                    self.data.$data.insert_tombstone(name.clone());
+                    self.subs.remove(rtype, name);
+                }
+            }};
+        }
+
+        match rtype {
+            ResourceType::Listener => tombstone_all!(listeners, rtype, names),
+            ResourceType::RouteConfiguration => tombstone_all!(route_configs, rtype, names),
+            ResourceType::Cluster => tombstone_all!(clusters, rtype, names),
+            ResourceType::ClusterLoadAssignment => tombstone_all!(load_assignments, rtype, names),
+        }
+    }
+
+    fn insert_listeners(
+        &mut self,
+        listeners: Vec<(ResourceVersion, xds_listener::Listener)>,
+    ) -> Vec<ResourceError> {
+        let mut errors = Vec::new();
+
+        for (version, listener) in listeners {
+            if !self.data.listeners.is_changed(&listener.name, &listener) {
+                continue;
+            }
+            let Some(sub) = self
+                .subs
+                .find_subcribed(ResourceType::Listener, &listener.name)
+            else {
+                continue;
+            };
+
+            let listener_name = listener.name.clone();
+            let api_listener = match ApiListener::from_xds(&listener_name, listener) {
+                Ok(l) => l,
+                Err(e) => {
+                    self.data
+                        .listeners
+                        .insert_error(listener_name, version.clone(), e.clone());
+                    errors.push(e);
+                    continue;
+                }
+            };
+
+            // reset all outgoing references
+            self.subs.reset_refs(sub);
+
+            match &api_listener.route_config {
+                // RDS: add a reference from this Listener to a RouteConfiguration
+                ApiListenerData::Rds(rc_name) => {
+                    self.subs
+                        .add_ref(sub, ResourceType::RouteConfiguration, rc_name.as_str());
+                }
+                // with an inline RouteConfiguration, add a reference to all of
+                // the clusters that this Listener points to.
+                ApiListenerData::Inlined(RouteConfigData::Route { clusters, .. }) => {
+                    for cluster in clusters {
+                        self.subs
+                            .add_ref(sub, ResourceType::Cluster, cluster.as_str());
+                    }
+                }
+                // policy RouteConfigurations are the *targets* of references,
+                // and shouldn't actually reference any resources themselves.
+                // the only thing we should do here is update the clusters that
+                // point at this policy.
+                ApiListenerData::Inlined(RouteConfigData::LbPolicy { action, cluster }) => {
+                    // inserting the cluster should have already subscribed to
+                    // this route implicitly! don't recreate the edge in the
+                    // other direction.
+                    //
+                    // rebuild the cluster with the new LbPolicy from this
+                    // listener. we don't have to use the ref graph here, since
+                    // we have the cluster name (effectively the parent-pointer)
+                    // in the LbPolicy.
+                    //
+                    // TODO: remove this from Listener and only have Routes serve this purpose.
+                    let version_and_xds = self.data.clusters.get(cluster.as_str()).and_then(|e| {
+                        let version = e.version();
+                        let data = e.data();
+                        version.zip(data).map(|(v, d)| (v.clone(), d.xds.clone()))
+                    });
+                    let res = match version_and_xds {
+                        Some((version, xds)) => {
+                            self.insert_cluster(version, xds, Some(Arc::clone(action)))
+                        }
+                        None => Ok(()),
+                    };
+                    if let Err(e) = res {
+                        self.data
+                            .listeners
+                            .insert_error(listener_name, version, e.clone());
+                        errors.push(e);
+                        continue;
+                    }
+                }
+            }
+
+            // do the update
+            self.subs
+                .clear_changes(ResourceType::Listener, &listener_name);
+            self.data
+                .listeners
+                .insert_ok(listener_name, version, api_listener);
+        }
+
+        errors
+    }
+
+    fn insert_clusters(
+        &mut self,
+        clusters: Vec<(ResourceVersion, xds_cluster::Cluster)>,
+    ) -> Vec<ResourceError> {
+        let mut errors = Vec::new();
+
+        for (version, cluster) in clusters {
+            if !self.data.clusters.is_changed(&cluster.name, &cluster) {
+                continue;
+            }
+
+            let lb_action = self.find_lb_action(&cluster.name);
+            if let Err(e) = self.insert_cluster(version, cluster, lb_action) {
+                errors.push(e);
+            }
+        }
+
+        errors
+    }
+
+    fn insert_cluster(
+        &mut self,
+        version: ResourceVersion,
+        cluster: xds_cluster::Cluster,
+        lb_policy: Option<Arc<xds_route::RouteAction>>,
+    ) -> Result<(), ResourceError> {
+        let Some(sub) = self
+            .subs
+            .find_subcribed(ResourceType::Cluster, &cluster.name)
+        else {
+            return Ok(());
+        };
+
+        let cluster_name = cluster.name.clone();
+        let cluster = match Cluster::from_xds(cluster, lb_policy.as_deref()) {
+            Ok(c) => c,
+            Err(e) => {
+                self.data
+                    .clusters
+                    .insert_error(cluster_name, version.clone(), e.clone());
+                return Err(e);
+            }
+        };
+
+        // reset all outgoing references since we know we're updating
+        self.subs.reset_refs(sub);
+
+        // point to the CLA for this cluster or start a DNS subscription for it.
+        match cluster.dns_name() {
+            Some(dns_name) => {
+                self.dns.add.insert(dns_name);
+            }
+            None => self
+                .subs
+                .add_ref(sub, ResourceType::ClusterLoadAssignment, &cluster_name),
+        }
+
+        // point to the LB config Listener for this cluster. pointing to the Listener
+        // means that the control plane has the option of sending us either a Listener
+        // or a RouteConfig for the Lb Config.
+        //
+        // TODO: make this a RouteConfig instead of a listener?
+        let lb_config_name = cluster.backend_lb.config.id.lb_config_route_name();
+        self.subs
+            .add_ref(sub, ResourceType::Listener, &lb_config_name);
+
+        // actually insert the data
+        self.subs
+            .clear_changes(ResourceType::Cluster, &cluster_name);
+        self.data.clusters.insert_ok(cluster_name, version, cluster);
+
+        Ok(())
+    }
+
+    fn find_lb_action(&self, cluster_name: &str) -> Option<Arc<xds_route::RouteAction>> {
+        let target = BackendId::from_str(cluster_name).ok()?;
+        let listener = self.data.listeners.get(&target.lb_config_route_name())?;
+
+        match &listener.data()?.route_config {
+            ApiListenerData::Rds(name) => {
+                let route_config = self.data.route_configs.get(name.as_str())?;
+                route_config.data().and_then(|rc| match &rc.data {
+                    RouteConfigData::LbPolicy { action, .. } => Some(action.clone()),
+                    _ => None,
+                })
+            }
+            ApiListenerData::Inlined(data) => match &data {
+                RouteConfigData::LbPolicy { action, .. } => Some(action.clone()),
+                _ => None,
+            },
+        }
+    }
+
+    fn insert_route_configs(
+        &mut self,
+        route_configs: Vec<(ResourceVersion, xds_route::RouteConfiguration)>,
+    ) -> Vec<ResourceError> {
+        let mut errors = Vec::new();
+
+        for (version, route_config) in route_configs {
+            let Some(sub) = self
+                .subs
+                .find_subcribed(ResourceType::RouteConfiguration, &route_config.name)
+            else {
+                continue;
+            };
+
+            let route_name = route_config.name.clone();
+            let route_config = match RouteConfig::from_xds(route_config) {
+                Ok(route_config) => route_config,
+                Err(e) => {
+                    self.data
+                        .route_configs
+                        .insert_error(route_name, version, e.clone());
+                    errors.push(e);
+                    continue;
+                }
+            };
+
+            match &route_config.data {
+                // add a ref to all downstream clusters
+                RouteConfigData::Route { clusters, .. } => {
+                    for cluster in clusters {
+                        self.subs
+                            .add_ref(sub, ResourceType::Cluster, cluster.as_str());
+                    }
+                }
+                // this is Lb policy, update the cluster it's attached to.
+                RouteConfigData::LbPolicy { action, cluster } => {
+                    // inserting the cluster should have already subscribed to
+                    // this route implicitly! don't recreate the edge in the
+                    // other direction.
+                    //
+                    // rebuild the cluster with the new LbPolicy from this
+                    // route. we don't have to use the ref graph here, since
+                    // we have the cluster name (effectively the parent-pointer)
+                    // in the LbPolicy.
+                    //
+                    // TODO: remove this from Listener and only have Routes serve this purpose.
+                    let version_and_xds = self.data.clusters.get(cluster.as_str()).and_then(|e| {
+                        let version = e.version();
+                        let data = e.data();
+                        version.zip(data).map(|(v, d)| (v.clone(), d.xds.clone()))
+                    });
+                    let res = match version_and_xds {
+                        Some((version, xds)) => {
+                            self.insert_cluster(version, xds, Some(Arc::clone(action)))
+                        }
+                        None => Ok(()),
+                    };
+                    if let Err(e) = res {
+                        self.data
+                            .route_configs
+                            .insert_error(route_name, version, e.clone());
+                        errors.push(e);
+                        continue;
+                    }
+                }
+            }
+
+            // complete the insert
+            self.subs
+                .clear_changes(ResourceType::RouteConfiguration, &route_name);
+            self.data
+                .route_configs
+                .insert_ok(route_name, version, route_config);
+        }
+
+        errors
+    }
+
+    fn insert_load_assignments(
+        &mut self,
+        load_assignments: Vec<(ResourceVersion, xds_endpoint::ClusterLoadAssignment)>,
+    ) -> Vec<ResourceError> {
+        let mut errors = Vec::new();
+
+        for (version, load_assignment) in load_assignments {
+            let sub = self.subs.find_subcribed(
+                ResourceType::ClusterLoadAssignment,
+                &load_assignment.cluster_name,
+            );
+            if sub.is_none() {
+                continue;
+            };
+
+            let cla_name = load_assignment.cluster_name.clone();
+            match LoadAssignment::from_xds(load_assignment) {
+                Ok(cla) => {
+                    self.subs
+                        .clear_changes(ResourceType::ClusterLoadAssignment, &cla_name);
+                    self.data.load_assignments.insert_ok(cla_name, version, cla);
+                }
+                Err(e) => {
+                    self.data
+                        .load_assignments
+                        .insert_error(cla_name, version, e.clone());
+                    errors.push(e);
+                }
+            };
+        }
+
+        errors
+    }
+}
+
+/// A read-only handle to a [Cache]. `CacheReader`s are cheap to clone and
+/// share.
 #[derive(Default, Clone)]
 pub(super) struct CacheReader {
     data: Arc<CacheData>,
 }
 
+impl ConfigCache for CacheReader {
+    async fn get_route<S: AsRef<str>>(&self, host: S) -> Option<Arc<Route>> {
+        let listener = self.data.listeners.get_await(host.as_ref()).await?;
+
+        match &listener.data()?.route_config {
+            ApiListenerData::Rds(name) => {
+                let route_config = self.data.route_configs.get_await(name.as_str()).await?;
+
+                match &route_config.data()?.data {
+                    RouteConfigData::Route { route, .. } => Some(route.clone()),
+                    _ => None,
+                }
+            }
+            ApiListenerData::Inlined(data) => match &data {
+                RouteConfigData::Route { route, .. } => Some(route.clone()),
+                _ => None,
+            },
+        }
+    }
+
+    async fn get_backend(&self, id: &BackendId) -> Option<Arc<BackendLb>> {
+        let cluster = self.data.clusters.get_await(&id.name()).await?;
+        let cluster_data = cluster.data()?;
+        Some(cluster_data.backend_lb.clone())
+    }
+
+    async fn get_endpoints(&self, id: &BackendId) -> Option<Arc<EndpointGroup>> {
+        let la = self.data.load_assignments.get_await(&id.name()).await?;
+        let la_data = la.data()?;
+        Some(la_data.endpoint_group.clone())
+    }
+}
+
 impl CacheReader {
+    /// Iterate over all routes currently in cache.
     pub(super) fn iter_routes(&self) -> impl Iterator<Item = Arc<Route>> + '_ {
         let listener_routes = self.data.listeners.iter().filter_map(|entry| {
             entry
@@ -350,6 +1131,7 @@ impl CacheReader {
         listener_routes.chain(route_config_routes)
     }
 
+    /// Iterate over all backends currently in cache.
     pub(super) fn iter_backends(&self) -> impl Iterator<Item = Arc<BackendLb>> + '_ {
         self.data
             .clusters
@@ -357,17 +1139,22 @@ impl CacheReader {
             .filter_map(|entry| entry.data().map(|cluster| cluster.backend_lb.clone()))
     }
 
+    /// Iterate over all xDS currently in cache.
     pub(super) fn iter_xds(&self) -> impl Iterator<Item = XdsConfig> + '_ {
+        use prost::Name;
+
         macro_rules! any_iter {
             ($field:ident, $xds_type:ty) => {
                 self.data.$field.iter().map(|entry| {
                     let name = entry.name().to_string();
                     let type_url = <$xds_type>::type_url();
-                    let version = Some(entry.version().clone());
+                    let version = entry.version().cloned();
+
                     let xds = entry.data().map(|data| {
                         protobuf::Any::from_msg(data.xds()).expect("generated invalid protobuf")
                     });
                     let last_error = entry.last_error().map(|(v, e)| (v.clone(), e.to_string()));
+
                     XdsConfig {
                         name,
                         type_url,
@@ -389,674 +1176,10 @@ impl CacheReader {
     }
 }
 
-impl ConfigCache for CacheReader {
-    async fn get_route<S: AsRef<str>>(&self, host: S) -> Option<Arc<Route>> {
-        let listener = self.data.listeners.get_await(host.as_ref()).await?;
-
-        match &listener.data()?.route_config {
-            ApiListenerData::Rds(name) => {
-                let route_config = self.data.route_configs.get_await(name.as_str()).await?;
-
-                route_config.data().and_then(|rc| match &rc.data {
-                    RouteConfigData::Route { route, .. } => Some(route.clone()),
-                    _ => None,
-                })
-            }
-            ApiListenerData::Inlined(data) => match &data {
-                RouteConfigData::Route { route, .. } => Some(route.clone()),
-                _ => None,
-            },
-        }
-    }
-
-    async fn get_backend(&self, id: &BackendId) -> Option<Arc<BackendLb>> {
-        let cluster = self.data.clusters.get_await(&id.name()).await?;
-        let cluster_data = cluster.data()?;
-        Some(cluster_data.backend_lb.clone())
-    }
-
-    async fn get_endpoints(&self, id: &BackendId) -> Option<Arc<EndpointGroup>> {
-        let la = self.data.load_assignments.get_await(&id.name()).await?;
-        let la_data = la.data()?;
-        Some(la_data.endpoint_group.clone())
-    }
-}
-
-/// Shared XDS and client configuration for a SotW XDS client.
-///
-/// A [Cache] is built on the fly by a single writer, with any number of
-/// [CacheReader]s providing read-only access. Readers do not necessarily get a
-/// consistent snapshot of configuration (XDS doesn't define what that might
-/// even mean!) - see [CacheReader] for a more thorough explanation.
-///
-/// A `Cache` handles tracking the current state of a XDS resources, validating
-/// and converting them to internal configuration, and tracking any references
-/// to other resource types.
-#[derive(Default, Debug)]
-pub(super) struct Cache {
-    refs: DiGraph<GCData, ()>,
-    data: Arc<CacheData>,
-}
-
-/// GC data tracked for every xDS resource.
-///
-/// A resource is `pinnned` if explicitly requested by a caller, and will never
-/// be removed from cache.
-///
-/// A resource is `deleted` if it was already pinned and was deleted because an
-/// xDS server or garbage collection removed it. Deleted resources will still
-/// be used when generating subscriptions.
-#[derive(Debug)]
-struct GCData {
-    name: String,
-    resource_type: ResourceType,
-    pinned: bool,
-    deleted: bool,
-    dns_name: Option<(Hostname, u16)>,
-}
-
-impl GCData {
-    fn is_gc_root(&self) -> bool {
-        self.pinned && !self.deleted
-    }
-}
-
-#[derive(Debug, Default)]
-struct CacheData {
-    listeners: ResourceMap<ApiListener>,
-    route_configs: ResourceMap<RouteConfig>,
-    clusters: ResourceMap<Cluster>,
-    load_assignments: ResourceMap<LoadAssignment>,
-}
-
-impl Cache {
-    /// Create a new read-only handle to this cache.
-    ///
-    /// Read handles are cheap, and intended to be created and shared across
-    /// multiple threads and tasks.
-    pub(super) fn reader(&self) -> CacheReader {
-        CacheReader {
-            data: self.data.clone(),
-        }
-    }
-
-    /// Explicitly subscribe to an XDS resource.
-    pub(super) fn subscribe(&mut self, resource_type: ResourceType, name: &str) -> bool {
-        let (node, created) = self.find_or_create_ref(resource_type, name);
-        self.pin_ref(node);
-        created
-    }
-
-    /// Return the resource names that should be subscribed to.
-    pub(super) fn subscriptions(&self, resource_type: ResourceType) -> Vec<String> {
-        let weights = self
-            .refs
-            .node_weights()
-            .filter(|n| n.resource_type == resource_type);
-        weights.map(|n| n.name.clone()).collect()
-    }
-
-    /// An iterator over all of the DNS names of external DNS clusters in the
-    /// cache.
-    pub(super) fn dns_names(&self) -> impl Iterator<Item = (Hostname, u16)> + '_ {
-        self.refs.node_weights().filter_map(|n| n.dns_name.clone())
-    }
-
-    /// Insert a batch of xDS resources into the cache.
-    ///
-    /// Returns the set of resource types that indicates the types of resources
-    /// that were changed or were newly subscribed to. Callers should use this
-    /// set to determine whether to ACK or to re-subscribe to resources of that
-    /// type.
-    pub(super) fn insert(
-        &mut self,
-        resources: ResourceVec,
-    ) -> (ResourceTypeSet, Vec<ResourceError>) {
-        let (changed, errs) = match resources {
-            ResourceVec::Listener(ls) => self.insert_listeners(ls),
-            ResourceVec::RouteConfiguration(rcs) => self.insert_route_configs(rcs),
-            ResourceVec::Cluster(cs) => self.insert_clusters(cs),
-            ResourceVec::ClusterLoadAssignment(clas) => self.insert_load_assignments(clas),
-        };
-
-        if !changed.is_empty() {
-            self.collect();
-        }
-
-        (changed, errs)
-    }
-
-    /// Unsubscribe from an XDS resource and explicitly delete it from cache.
-    pub(super) fn delete(&mut self, resource_type: ResourceType, name: &str) -> bool {
-        if !self.delete_ref(resource_type, name, true) {
-            return false;
-        }
-
-        match resource_type {
-            ResourceType::Cluster => {
-                self.data.clusters.remove(name);
-            }
-            ResourceType::ClusterLoadAssignment => {
-                self.data.load_assignments.remove(name);
-            }
-            ResourceType::Listener => {
-                self.data.listeners.remove(name);
-            }
-            ResourceType::RouteConfiguration => {
-                self.data.route_configs.remove(name);
-            }
-        }
-
-        true
-    }
-}
-
-/// safety: `petgraph` NodeIndexes are unstable - when removing a node, it may
-/// invalidate an index we previously looked up. It is not sound to hold on to
-/// an index from before a deletion took place.
-///
-/// In practice, this means DO NOT save NodeIndexes anywhere. Use them as locals
-/// but don't store them in struct fields etc. This makes it impractical to build
-/// e.g. a lookup from (ResourceType, Name) -> NodeIndex. If that becomes necessary
-/// petgraph is working on a StableGraph where NodeIndexes are never invalidated.
-impl Cache {
-    fn collect(&mut self) {
-        use visit::{Control, DfsEvent};
-
-        // walk the GC graph, keeping the set of the reachable nodes.
-        //
-        // lean on petgraph's Control to only visit each node once - because
-        // the ref graph must be a DAG, we can skip marking nodes twice and
-        // emit Control::Prune every time we see a node we've already seen.
-        let mut reachable = self.refs.visit_map();
-        visit::depth_first_search(&self.refs, self.gc_roots(), |event| -> Control<()> {
-            if let DfsEvent::Discover(n, _) = event {
-                if reachable.contains(n.index()) {
-                    return Control::Prune;
-                }
-                reachable.insert(n.index());
-            };
-
-            Control::Continue
-        });
-
-        let unreachable_nodes = self
-            .refs
-            .node_indices()
-            .filter(|n| !reachable.contains(n.index()));
-
-        let mut unreachable_names: EnumMap<ResourceType, Vec<String>> = EnumMap::default();
-        for n in unreachable_nodes {
-            let n = &self.refs[n];
-            unreachable_names[n.resource_type].push(n.name.to_string());
-        }
-
-        for (resource_type, names) in unreachable_names.into_iter() {
-            match resource_type {
-                ResourceType::Listener => self.data.listeners.remove_all(&names),
-                ResourceType::RouteConfiguration => self.data.route_configs.remove_all(&names),
-                ResourceType::Cluster => self.data.clusters.remove_all(&names),
-                ResourceType::ClusterLoadAssignment => {
-                    self.data.load_assignments.remove_all(&names);
-                }
-            }
-        }
-
-        // safety: no longer holding any NodeIndexes, it's safe to invalidate
-        // any outstanding ref by calling retain_nodes
-        self.refs
-            .retain_nodes(|g, n| g[n].pinned || reachable.contains(n.index()));
-    }
-
-    fn insert_listeners(
-        &mut self,
-        listeners: Vec<(ResourceVersion, xds_listener::Listener)>,
-    ) -> (ResourceTypeSet, Vec<ResourceError>) {
-        let mut changed = ResourceTypeSet::default();
-        let mut errors = Vec::new();
-        let mut to_remove: BTreeSet<_> = self.data.listeners.names().collect();
-
-        for (version, listener) in listeners {
-            to_remove.remove(&listener.name);
-
-            if self.data.listeners.is_changed(&listener.name, &listener) {
-                let listener_name = listener.name.clone();
-                let api_listener = match ApiListener::from_xds(&listener_name, listener) {
-                    Ok(l) => l,
-                    Err(e) => {
-                        self.data
-                            .listeners
-                            .insert_error(listener_name, version.clone(), e.clone());
-                        errors.push(e);
-                        continue;
-                    }
-                };
-
-                // remove the downstream route config ref and replace it with a new one
-                let (node, _) = self.find_or_create_ref(ResourceType::Listener, &listener_name);
-                self.reset_ref(node);
-
-                match &api_listener.route_config {
-                    ApiListenerData::Rds(name) => {
-                        // for RDS, update refs so we have new xds subscriptions
-
-                        let (rc_node, created) = self
-                            .find_or_create_ref(ResourceType::RouteConfiguration, name.as_str());
-                        self.refs.update_edge(node, rc_node, ());
-
-                        if created {
-                            changed.insert(ResourceType::RouteConfiguration);
-                        }
-                    }
-                    ApiListenerData::Inlined(RouteConfigData::Route { clusters, .. }) => {
-                        // for inline RDS, update cluster refs for everything downstream
-
-                        let mut clusters_changed = false;
-
-                        for cluster in clusters {
-                            let (cluster_node, created) =
-                                self.find_or_create_ref(ResourceType::Cluster, cluster.as_str());
-                            self.refs.update_edge(node, cluster_node, ());
-                            clusters_changed |= created;
-                        }
-
-                        if clusters_changed {
-                            changed.insert(ResourceType::Cluster);
-                        }
-                    }
-                    ApiListenerData::Inlined(RouteConfigData::LbPolicy { action, cluster }) => {
-                        // for an LB policy update, recompute the lb config for
-                        // the cluster. have to set the gc graph subscription in
-                        // case this cluster doesn't exist yet.
-                        let (cluster_node, created) =
-                            self.find_or_create_ref(ResourceType::Cluster, cluster.as_str());
-                        self.refs.update_edge(node, cluster_node, ());
-                        if created {
-                            changed.insert(ResourceType::Cluster);
-                        }
-
-                        if let Err(e) = self.rebuild_cluster(&mut changed, cluster, action) {
-                            self.data.listeners.insert_error(
-                                listener_name,
-                                version.clone(),
-                                e.clone(),
-                            );
-                            errors.push(e);
-                            continue;
-                        }
-                    }
-                }
-
-                // insert data into cache
-                self.data
-                    .listeners
-                    .insert_ok(listener_name, version.clone(), api_listener);
-                changed.insert(ResourceType::Listener);
-            }
-        }
-
-        // safety: the refs graph should be in sync with the names of clusters,
-        // so panic here if we try to remove a name that didn't exist in the ref
-        // graph.
-        //
-        // this guarantee comes from there only being a single cache writer.
-        for name in to_remove {
-            changed.insert(ResourceType::Listener);
-            self.delete_ref(ResourceType::Listener, &name, false);
-            self.data.listeners.remove(&name);
-        }
-
-        (changed, errors)
-    }
-
-    fn insert_clusters(
-        &mut self,
-        clusters: Vec<(ResourceVersion, xds_cluster::Cluster)>,
-    ) -> (ResourceTypeSet, Vec<ResourceError>) {
-        let mut changed = ResourceTypeSet::default();
-        let mut errors = Vec::new();
-        let mut to_remove: BTreeSet<_> = self.data.clusters.names().collect();
-
-        for (version, cluster) in clusters {
-            to_remove.remove(&cluster.name);
-
-            if self.data.clusters.is_changed(&cluster.name, &cluster) {
-                let lb_action = self.find_lb_action(&cluster.name);
-                if let Err(e) =
-                    self.insert_cluster(&mut changed, &version, cluster, lb_action.as_deref())
-                {
-                    errors.push(e);
-                }
-            }
-        }
-
-        // safety: the refs graph should be in sync with the names of clusters,
-        // so panic here if we try to remove a name that didn't exist in the ref
-        // graph.
-        //
-        // this guarantee comes from there only being a single cache writer.
-        for name in to_remove {
-            changed.insert(ResourceType::Cluster);
-            self.delete_ref(ResourceType::Cluster, &name, false);
-            self.data.clusters.remove(&name);
-        }
-
-        (changed, errors)
-    }
-
-    fn insert_route_configs(
-        &mut self,
-        route_configs: Vec<(ResourceVersion, xds_route::RouteConfiguration)>,
-    ) -> (ResourceTypeSet, Vec<ResourceError>) {
-        let mut errors = Vec::new();
-        let mut changed = ResourceTypeSet::default();
-
-        for (version, route_config) in route_configs {
-            if self
-                .data
-                .route_configs
-                .is_changed(&route_config.name, &route_config)
-            {
-                // we need the ref here to set up edges to clusters. if we
-                // haven't subscribed to this CLA, don't bother creating a new
-                // ref - it'll just get GCd anyway.
-                let Some(cla_ref) =
-                    self.find_ref(ResourceType::RouteConfiguration, &route_config.name)
-                else {
-                    continue;
-                };
-
-                let route_config_name = route_config.name.clone();
-                let route_config = match RouteConfig::from_xds(route_config) {
-                    Ok(rc) => rc,
-                    Err(e) => {
-                        self.data.route_configs.insert_error(
-                            route_config_name,
-                            version.clone(),
-                            e.clone(),
-                        );
-                        errors.push(e);
-                        continue;
-                    }
-                };
-
-                match &route_config.data {
-                    RouteConfigData::Route { clusters, .. } => {
-                        // update the reference graph to include edges from this route to clusters.
-                        self.reset_ref(cla_ref);
-                        for cluster in clusters {
-                            let (cluster_node, created) =
-                                self.find_or_create_ref(ResourceType::Cluster, cluster.as_str());
-                            self.refs.update_edge(cla_ref, cluster_node, ());
-
-                            if created {
-                                changed.insert(ResourceType::Cluster);
-                            }
-                        }
-                    }
-                    RouteConfigData::LbPolicy { action, cluster } => {
-                        // if this looks like the default RouteConfiguration for
-                        // a Cluster, rebuild it. don't track resource type
-                        // changes here, since rebuild_cluster will also track
-                        // changes.
-                        let (cluster_node, _) =
-                            self.find_or_create_ref(ResourceType::Cluster, cluster.as_str());
-                        self.refs.update_edge(cla_ref, cluster_node, ());
-
-                        if let Err(e) = self.rebuild_cluster(&mut changed, cluster, action) {
-                            errors.push(e);
-                        }
-                    }
-                }
-
-                // actually insert the route config
-                self.data
-                    .route_configs
-                    .insert_ok(route_config_name, version.clone(), route_config);
-                changed.insert(ResourceType::RouteConfiguration);
-            }
-        }
-
-        (changed, errors)
-    }
-
-    fn insert_load_assignments(
-        &mut self,
-        load_assignments: Vec<(ResourceVersion, xds_endpoint::ClusterLoadAssignment)>,
-    ) -> (ResourceTypeSet, Vec<ResourceError>) {
-        use ResourceType::*;
-
-        let mut changed = ResourceTypeSet::default();
-        let mut errors = Vec::new();
-
-        for (version, load_assignment) in load_assignments {
-            if self
-                .data
-                .load_assignments
-                .is_changed(&load_assignment.cluster_name, &load_assignment)
-            {
-                // ignore this CLA if the cache is not subscribed
-                if self
-                    .find_ref(ClusterLoadAssignment, &load_assignment.cluster_name)
-                    .is_none()
-                {
-                    continue;
-                }
-
-                let load_assignment_name = load_assignment.cluster_name.clone();
-                let load_assignment = match LoadAssignment::from_xds(load_assignment) {
-                    Ok(cla) => cla,
-                    Err(e) => {
-                        self.data.load_assignments.insert_error(
-                            load_assignment_name.clone(),
-                            version.clone(),
-                            e.clone(),
-                        );
-                        errors.push(e);
-                        continue;
-                    }
-                };
-
-                self.data.load_assignments.insert_ok(
-                    load_assignment_name,
-                    version.clone(),
-                    load_assignment,
-                );
-                changed.insert(ClusterLoadAssignment);
-            }
-        }
-
-        (changed, errors)
-    }
-
-    /// Try to rebuild a Cluster and its data, using a new RouteAction to fill
-    /// in its load balancing policies.
-    fn rebuild_cluster(
-        &mut self,
-        changed: &mut ResourceTypeSet,
-        cluster: &ResourceName<Cluster>,
-        route_action: &xds_route::RouteAction,
-    ) -> Result<(), ResourceError> {
-        // clone the cluster's current version and xds so that borrowck doesn't
-        // get mad. it gets upset about a partial borrow of cache data while
-        // we're trying to mutate the ref graph in build_cluster.
-        //
-        // this is a relatively cheap clone so whatever.
-        let version_and_xds = self.data.clusters.get(cluster.as_str()).and_then(|e| {
-            let version = e.version();
-            e.data().map(|d| (version.clone(), d.xds().clone()))
-        });
-
-        match version_and_xds {
-            Some((version, xds)) => self.insert_cluster(changed, &version, xds, Some(route_action)),
-            None => Ok(()),
-        }
-    }
-
-    /// Build and insert a Cluster based on new xDS, update it's GC refs, etc.
-    /// This also has the side effect of inserting a subscription for this
-    /// Cluster's default Listener into the cluster.
-    ///
-    /// This is split out from the inner loop of `insert_clusters` so that it
-    /// can be called whenever RouteConfigurations or Listeners with default
-    /// routing info change.
-    fn insert_cluster(
-        &mut self,
-        changed: &mut ResourceTypeSet,
-        version: &crate::xds::ResourceVersion,
-        cluster: xds_cluster::Cluster,
-        default_action: Option<&xds_route::RouteAction>,
-    ) -> Result<(), ResourceError> {
-        // try to find this cluster in the ref graph. it's possible it's
-        // now from a stale subscription.
-        let Some(node) = self.find_ref(ResourceType::Cluster, &cluster.name) else {
-            return Ok(());
-        };
-
-        let cluster_name = cluster.name.clone();
-        let cluster = match Cluster::from_xds(cluster, default_action) {
-            Ok(c) => c,
-            Err(e) => {
-                self.data
-                    .clusters
-                    .insert_error(cluster_name, version.clone(), e.clone());
-                return Err(e);
-            }
-        };
-
-        // clear the outgoing edges form this node.
-        self.reset_ref(node);
-
-        // remove the old CLA edge and replace it with a new one.
-        changed.insert(ResourceType::ClusterLoadAssignment);
-        let (cla_node, _) =
-            self.find_or_create_ref(ResourceType::ClusterLoadAssignment, &cluster_name);
-        self.refs.update_edge(node, cla_node, ());
-
-        // try to subscribe to the lb_config Listener for this Cluster if it
-        // doesn't already exist in the GC graph.
-        let lb_config_listener_name = cluster.backend_lb.config.id.lb_config_route_name();
-        let (default_listener_node, created) =
-            self.find_or_create_ref(ResourceType::Listener, &lb_config_listener_name);
-        if created {
-            changed.insert(ResourceType::Listener);
-        }
-        self.refs.update_edge(node, default_listener_node, ());
-
-        // insert the cluster
-        self.data
-            .clusters
-            .insert_ok(cluster_name, version.clone(), cluster);
-        changed.insert(ResourceType::Cluster);
-
-        Ok(())
-    }
-
-    /// Return the node indices of all GC roots. GC roots are either Listeners
-    /// or are explicitly pinned.
-    ///
-    /// Safety: NodeIndexes are not stable across deletions. This vec is not
-    /// safe to store long term or between collections.
-    fn gc_roots(&self) -> Vec<NodeIndex> {
-        self.refs
-            .node_indices()
-            .filter(|idx| self.refs[*idx].is_gc_root())
-            .collect()
-    }
-
-    /// Delete a ref from the GC map.
-    fn delete_ref(&mut self, resource_type: ResourceType, name: &str, force: bool) -> bool {
-        match self.find_ref(resource_type, name) {
-            Some(node) => {
-                if force || !self.refs[node].pinned {
-                    self.refs.remove_node(node);
-                    true
-                } else {
-                    self.refs[node].deleted = true;
-                    false
-                }
-            }
-            None => false,
-        }
-    }
-
-    fn find_lb_action(&self, cluster_name: &str) -> Option<Arc<xds_route::RouteAction>> {
-        let target = BackendId::from_str(cluster_name).ok()?;
-        let listener = self.data.listeners.get(&target.lb_config_route_name())?;
-
-        match &listener.data()?.route_config {
-            ApiListenerData::Rds(name) => {
-                let route_config = self.data.route_configs.get(name.as_str())?;
-                route_config.data().and_then(|rc| match &rc.data {
-                    RouteConfigData::LbPolicy { action, .. } => Some(action.clone()),
-                    _ => None,
-                })
-            }
-            ApiListenerData::Inlined(data) => match &data {
-                RouteConfigData::LbPolicy { action, .. } => Some(action.clone()),
-                _ => None,
-            },
-        }
-    }
-
-    /// Find or create a GC ref for the given name or resource type.
-    ///
-    /// New GC refs are not marked `reachable` by default.
-    fn find_or_create_ref(&mut self, resource_type: ResourceType, name: &str) -> (NodeIndex, bool) {
-        if let Some(node) = self.find_ref(resource_type, name) {
-            self.refs[node].deleted = false;
-            return (node, false);
-        }
-
-        let mut dns_name = None;
-        if resource_type == ResourceType::Cluster {
-            if let Ok(backend) = BackendId::from_str(name) {
-                if let Service::Dns(dns) = backend.service {
-                    dns_name = Some((dns.hostname, backend.port))
-                }
-            }
-        }
-
-        let node = self.refs.add_node(GCData {
-            name: name.to_string(),
-            resource_type,
-            pinned: false,
-            deleted: false,
-            dns_name,
-        });
-        (node, true)
-    }
-
-    /// Find a GC ref with the given resource type or name.
-    fn find_ref(&self, resource_type: ResourceType, name: &str) -> Option<NodeIndex> {
-        self.refs.node_indices().find(|n| {
-            let n = &self.refs[*n];
-            n.resource_type == resource_type && n.name == name
-        })
-    }
-
-    fn pin_ref(&mut self, node: NodeIndex) {
-        self.refs[node].pinned = true;
-    }
-
-    /// Remove all of a GC ref's outgoing edges.
-    fn reset_ref(&mut self, node: NodeIndex) {
-        let neighbors: Vec<_> = self
-            .refs
-            .neighbors_directed(node, Direction::Outgoing)
-            .collect();
-
-        for n in neighbors {
-            if let Some((edge, _)) = self.refs.find_edge_undirected(node, n) {
-                self.refs.remove_edge(edge);
-            };
-        }
-    }
-}
-
 #[cfg(test)]
 mod test {
-    use junction_api::{backend::LbPolicy, Service};
+    use junction_api::Service;
+    use pretty_assertions::assert_eq;
 
     use super::*;
     use crate::xds::test as xds_test;
@@ -1065,937 +1188,756 @@ mod test {
     fn assert_sync<T: Sync>() {}
 
     #[test]
-    fn assert_reader_send_sync() {
+    fn test_reader_send_sync() {
         assert_send::<CacheReader>();
         assert_sync::<CacheReader>();
     }
 
-    #[track_caller]
-    fn assert_insert((changed, errors): (ResourceTypeSet, Vec<ResourceError>)) -> ResourceTypeSet {
-        assert!(errors.is_empty(), "{errors:?}");
-        changed
+    #[test]
+    fn test_cache_send_sync() {
+        assert_send::<Cache>();
+        assert_sync::<Cache>();
     }
 
-    #[track_caller]
-    fn assert_subscribe_insert(cache: &mut Cache, resources: ResourceVec) -> ResourceTypeSet {
-        for name in resources.names() {
-            cache.subscribe(resources.resource_type(), &name);
+    macro_rules! collect_str {
+        ($($arg:expr),* $(,)?) => {
+            [$(
+                    $arg.to_string(),
+            )*].into_iter().collect()
         }
-        assert_insert(cache.insert(resources))
+    }
+
+    macro_rules! collect_kv_str {
+        ($(($k:expr, $v:expr)),* $(,)?) => {
+            [$(
+                ($k.to_string(), $v.to_string()),
+            )*].into_iter().collect()
+        }
     }
 
     #[track_caller]
-    fn assert_changed(changed: &ResourceTypeSet, expected: impl IntoIterator<Item = ResourceType>) {
-        let mut actual: Vec<_> = changed.values().collect();
-        actual.sort();
-
-        let mut expected: Vec<_> = expected.into_iter().collect();
-        expected.sort();
-
-        assert_eq!(actual, expected);
+    fn assert_insert(errors: Vec<ResourceError>) {
+        assert!(errors.is_empty(), "errors is not empty: {errors:?}");
     }
 
     #[test]
-    fn test_insert_invalid_listener() {
+    fn test_empty_cache() {
         let mut cache = Cache::default();
 
-        // insert a listener with no api_listener
+        // changes should return nothing
+        let (rs, dns) = cache.collect();
+        assert!(rs.values().all(|v| v.is_empty()));
+        assert!(dns.is_noop());
+
+        // there should be no initial versions
+        assert!(ResourceType::all()
+            .iter()
+            .all(|&rtype| cache.versions(rtype).is_empty()));
+
+        // there should be no subscriptions
+        assert!(cache.subscriptions(ResourceType::Listener).is_empty());
+    }
+
+    #[test]
+    fn test_insert_listener_lds_explicit() {
+        let listeners = ResourceVec::from_listeners(
+            "123".into(),
+            vec![xds_test::listener!(
+                "listener.example.svc.cluster.local",
+                "example-route",
+            )],
+        );
+
+        let mut cache = Cache::default();
+        cache.set_wildcard(ResourceType::Listener, false);
+
+        // insert with no errors, and no effects
+        assert_insert(cache.insert(listeners.clone()));
+        let (resources, dns) = cache.collect();
+        assert_eq!(resources, EnumMap::default());
+        assert!(dns.is_noop());
+
+        // subscribe and clear the resulting subs for the listener
+        cache.subscribe(ResourceType::Listener, "listener.example.svc.cluster.local");
+        let _ = cache.collect();
+
+        // insert with no errors and generate the subscription to the cluster
+        assert_insert(cache.insert(listeners));
+        let (resources, dns) = cache.collect();
+        assert!(dns.is_noop());
+        assert_eq!(
+            resources,
+            enum_map::enum_map! {
+                ResourceType::RouteConfiguration => Changes {
+                    added: collect_str!["example-route"],
+                    removed: BTreeSet::new(),
+                 },
+                _ => Changes::default(),
+            }
+        );
+
+        // listener subscriptions should be the listener subscribed to versions
+        // and should report the version of the listener we have.
+        assert_eq!(
+            cache.subscriptions(ResourceType::Listener),
+            vec!["listener.example.svc.cluster.local"]
+        );
+        assert_eq!(
+            cache.versions(ResourceType::Listener),
+            collect_kv_str![("listener.example.svc.cluster.local", "123")]
+        );
+    }
+
+    #[test]
+    fn test_insert_listener_lds_wildcard() {
+        let mut cache = Cache::default();
+
+        assert_insert(cache.insert(ResourceVec::from_listeners(
+            "123".into(),
+            vec![xds_test::listener!(
+                "listener.example.svc.cluster.local",
+                "example-route",
+            )],
+        )));
+
+        // check that we've added an explicit subscription to the new cluster
+        // and that there are no DNS updates.
+        let (resources, dns) = cache.collect();
+        assert!(dns.is_noop());
+        assert_eq!(
+            resources,
+            enum_map::enum_map! {
+                ResourceType::RouteConfiguration => Changes {
+                    added: collect_str!["example-route"],
+                    removed: BTreeSet::new(),
+                 },
+                _ => Changes::default(),
+            }
+        );
+
+        // the listener subscription list should be empty - this listener is a
+        // wildcard resource. the current versions for listener/cluster should
+        // be just the listener in cache
+        assert!(cache.subscriptions(ResourceType::Listener).is_empty());
+        assert_eq!(
+            cache.versions(ResourceType::Listener),
+            collect_kv_str![("listener.example.svc.cluster.local", "123")]
+        );
+    }
+
+    #[test]
+    fn test_insert_listener_lds_inline_rds() {
+        let mut cache = Cache::default();
+
+        assert_insert(cache.insert(ResourceVec::from_listeners(
+            "123".into(),
+            vec![xds_test::listener!(
+                "listener.example.svc.cluster.local:80",
+                "example-route" => [xds_test::vhost!(
+                    "a-virtual-host",
+                    ["listener.example.svc.cluster.local"],
+                    [xds_test::route!(default "cluster.example:8008")],
+                )],
+            )],
+        )));
+
+        // check that we've added an explicit subscription to the new cluster
+        // and that there are no DNS updates.
+        let (resources, dns) = cache.collect();
+        assert!(dns.is_noop());
+        assert_eq!(
+            resources,
+            enum_map::enum_map! {
+                ResourceType::Cluster => Changes {
+                    added: collect_str!["cluster.example:8008"],
+                    removed: BTreeSet::new(),
+                 },
+                _ => Changes::default(),
+            }
+        );
+    }
+
+    #[test]
+    fn test_insert_listener_invalid() {
+        let mut cache = Cache::default();
         cache.subscribe(ResourceType::Listener, "potato");
-        let (changed, errors) = cache.insert(ResourceVec::from_listeners(
+        // clear subscription changes
+        let _ = cache.collect();
+
+        // the invalid insert should return an error
+        let errors = cache.insert(ResourceVec::from_listeners(
             "123".into(),
             [xds_listener::Listener {
                 name: "potato".to_string(),
                 ..Default::default()
             }],
         ));
-
-        assert!(changed.is_empty());
         assert_eq!(errors.len(), 1);
 
-        let listener_data = cache.data.listeners.get("potato").unwrap();
-        assert!(listener_data.data().is_none());
-        assert!(listener_data.name() == "potato");
-        assert!(*listener_data.version() == "".into());
-        assert!(matches!(listener_data.last_error(), Some((v, _)) if *v == "123".into()));
+        // should not have changed the cache
+        assert_eq!(cache.subscriptions(ResourceType::Listener), vec!["potato"]);
+        assert!(cache.versions(ResourceType::Listener).is_empty());
+        let (resources, dns) = cache.collect();
+        assert_eq!(resources, Default::default());
+        assert!(dns.is_noop());
     }
 
     #[test]
-    fn test_insert_listener_inline_route_config() {
+    fn test_insert_cluster_cds_wildcard() {
         let mut cache = Cache::default();
 
-        assert_subscribe_insert(
-            &mut cache,
-            ResourceVec::from_listeners(
-                "123".into(),
-                [xds_test::listener!(
-                    "listener.example.svc.cluster.local:80", "example-route" => [xds_test::vhost!(
-                        "vhost1.example.svc.cluster.local",
-                        ["listener.example.svc.cluster.local"],
-                        [xds_test::route!(default "cluster.example:80")],
-                    )],
-                )],
-            ),
-        );
+        let kube_backend = BackendId {
+            service: Service::kube("default", "whatever").unwrap(),
+            port: 7890,
+        };
+        let dns_backend = BackendId {
+            service: Service::dns("cluster.example").unwrap(),
+            port: 4433,
+        };
 
-        assert!(cache
-            .data
-            .listeners
-            .get("listener.example.svc.cluster.local:80")
-            .is_some());
-        assert!(cache.data.route_configs.is_empty());
-    }
-
-    #[test]
-    fn test_insert_listener_rds() {
-        let mut cache = Cache::default();
-
-        // insert two Listeners, should have changes and subscriptions for
-        // both LDS and RDS now.
-        let changed = assert_subscribe_insert(
-            &mut cache,
-            ResourceVec::from_listeners(
-                "123".into(),
-                [
-                    xds_test::listener!(
-                        "listener1.example.svc.cluster.local:8080",
-                        "example-route-1",
-                    ),
-                    xds_test::listener!(
-                        "listener2.example.svc.cluster.local:8080",
-                        "example-route-2",
-                    ),
-                ],
-            ),
-        );
-        assert_changed(
-            &changed,
-            [ResourceType::Listener, ResourceType::RouteConfiguration],
-        );
-        assert_eq!(
-            cache.data.listeners.names().collect::<Vec<_>>(),
+        assert_insert(cache.insert(ResourceVec::from_clusters(
+            "123".into(),
             vec![
-                "listener1.example.svc.cluster.local:8080",
-                "listener2.example.svc.cluster.local:8080"
+                xds_test::cluster!(dns_backend.name().leak()),
+                xds_test::cluster!(kube_backend.name().leak()),
             ],
-        );
-        assert!(cache.data.route_configs.is_empty());
+        )));
+
+        let (resources, dns) = cache.collect();
         assert_eq!(
-            cache.subscriptions(ResourceType::RouteConfiguration),
-            vec!["example-route-1", "example-route-2"],
+            resources,
+            enum_map::enum_map! {
+                ResourceType::Listener => Changes {
+                    added: collect_str!(
+                        kube_backend.lb_config_route_name(),
+                        dns_backend.lb_config_route_name(),
+                    ),
+                    ..Default::default()
+                },
+                ResourceType::ClusterLoadAssignment => Changes {
+                    added: collect_str![kube_backend.name()],
+                    ..Default::default()
+                 },
+                _ => Changes::default(),
+            }
         );
+        assert_eq!(
+            dns,
+            DnsUpdates {
+                add: BTreeSet::from_iter([(Hostname::from_static("cluster.example"), 4433)]),
+                ..Default::default()
+            },
+        );
+
+        // subscriptions should be empty but current versions should match
         assert!(cache.subscriptions(ResourceType::Cluster).is_empty());
-
-        // insert a RouteConfiguration. should have RDS and CDS changes and
-        // subscriptions after insert.
-        let changed = assert_insert(cache.insert(ResourceVec::from_route_configs(
-            "123".into(),
-            [xds_test::route_config!(
-                "example-route-1",
-                [xds_test::vhost!(
-                    "listener1.example.svc.cluster.local",
-                    ["listener2.example.svc.cluster.local"],
-                    [xds_test::route!(default "cluster1.example:8913")],
-                )]
-            )],
-        )));
-
-        assert_changed(
-            &changed,
-            [ResourceType::RouteConfiguration, ResourceType::Cluster],
-        );
         assert_eq!(
-            cache.data.listeners.names().collect::<Vec<_>>(),
-            vec![
-                "listener1.example.svc.cluster.local:8080",
-                "listener2.example.svc.cluster.local:8080"
-            ],
-        );
-        assert_eq!(
-            cache.data.route_configs.names().collect::<Vec<_>>(),
-            vec!["example-route-1"],
-        );
-        assert_eq!(
-            cache.subscriptions(ResourceType::RouteConfiguration),
-            vec!["example-route-1", "example-route-2"],
-        );
-        assert_eq!(
-            cache.subscriptions(ResourceType::Cluster),
-            vec!["cluster1.example:8913"],
-        );
-
-        let changed = assert_insert(cache.insert(ResourceVec::from_route_configs(
-            "123".into(),
-            [xds_test::route_config!(
-                "example-route-2",
-                [xds_test::vhost!(
-                    "listener2.example.svc.cluster.local",
-                    ["listener2.example.svc.cluster.local"],
-                    [xds_test::route!(default "cluster1.example:8913")],
-                )]
-            )],
-        )));
-        assert_changed(&changed, [ResourceType::RouteConfiguration]);
-
-        assert_eq!(
-            cache.data.listeners.names().collect::<Vec<_>>(),
-            vec![
-                "listener1.example.svc.cluster.local:8080",
-                "listener2.example.svc.cluster.local:8080"
-            ],
-        );
-        assert_eq!(
-            cache.data.route_configs.names().collect::<Vec<_>>(),
-            vec!["example-route-1", "example-route-2"],
-        );
-        assert_eq!(
-            cache.subscriptions(ResourceType::Cluster),
-            vec!["cluster1.example:8913"],
+            cache.versions(ResourceType::Cluster),
+            collect_kv_str![(kube_backend.name(), "123"), (dns_backend.name(), "123"),]
         );
     }
 
     #[test]
-    fn test_insert_cluster_eds() {
+    fn test_insert_cluster_cds_explicit() {
         let mut cache = Cache::default();
+        cache.set_wildcard(ResourceType::Cluster, false);
 
-        assert_subscribe_insert(
-            &mut cache,
-            ResourceVec::from_clusters("123".into(), [xds_test::cluster!("cluster1.example:8913")]),
+        let kube_backend = BackendId {
+            service: Service::kube("default", "whatever").unwrap(),
+            port: 7890,
+        };
+        let dns_backend = BackendId {
+            service: Service::dns("cluster.example").unwrap(),
+            port: 4433,
+        };
+
+        // subscribe only to the kube backend, clear changes
+        cache.subscribe(ResourceType::Cluster, &kube_backend.name());
+        let _ = cache.collect();
+
+        // insert both clusters at the same version
+        assert_insert(cache.insert(ResourceVec::from_clusters(
+            "123".into(),
+            vec![
+                xds_test::cluster!(dns_backend.name().leak()),
+                xds_test::cluster!(kube_backend.name().leak()),
+            ],
+        )));
+
+        // only the kbue cluster should have had an effect, no DNS updates
+        let (resources, dns) = cache.collect();
+        assert_eq!(
+            resources,
+            enum_map::enum_map! {
+                ResourceType::Listener => Changes {
+                    added: collect_str!(
+                        kube_backend.lb_config_route_name(),
+                    ),
+                    ..Default::default()
+                },
+                ResourceType::ClusterLoadAssignment => Changes {
+                    added: collect_str![kube_backend.name()],
+                    ..Default::default()
+                 },
+                _ => Changes::default(),
+            }
+        );
+        assert!(dns.is_noop());
+
+        // should have the explicit subscription to one cluster and one version
+        assert_eq!(
+            cache.subscriptions(ResourceType::Cluster),
+            vec![kube_backend.name()],
+        );
+        assert_eq!(
+            cache.versions(ResourceType::Cluster),
+            collect_kv_str![(kube_backend.name(), "123")]
+        );
+    }
+
+    #[test]
+    fn test_insert_route_config() {
+        let route_config = xds_test::route_config!(
+            "example-route",
+            vec![xds_test::vhost!(
+                "a-vhost",
+                ["listener.example.svc.cluster.local"],
+                [xds_test::route!(default "cluster.example:8008")]
+            )]
         );
 
-        assert!(cache.data.listeners.is_empty());
+        let mut cache = Cache::default();
+
+        // inserting with no subscription is empty
+        assert_insert(cache.insert(ResourceVec::from_route_configs(
+            "123".into(),
+            vec![route_config.clone()],
+        )));
+        let (resources, dns) = cache.collect();
+        assert!(resources.values().all(|c| c.is_empty()));
+        assert!(dns.is_noop());
         assert!(cache.data.route_configs.is_empty());
-        assert!(cache.data.clusters.get("cluster1.example:8913").is_some());
-        assert!(cache.data.load_assignments.is_empty());
+
+        // insert listener, should now be able to insert the route config
+        assert_insert(cache.insert(ResourceVec::from_listeners(
+            "123".into(),
+            vec![xds_test::listener!(
+                "listener.example.svc.cluster.local",
+                "example-route"
+            )],
+        )));
+
+        // should now have a new reference to a cluster
+        assert_insert(cache.insert(ResourceVec::from_route_configs(
+            "123".into(),
+            vec![route_config],
+        )));
+        let (resources, dns) = cache.collect();
+        assert_eq!(
+            resources,
+            enum_map::enum_map! {
+                ResourceType::Cluster => Changes {
+                    added: collect_str!["cluster.example:8008"],
+                    ..Default::default()
+                },
+                _ => Changes::default(),
+            }
+        );
+        assert!(dns.is_noop());
+
+        assert_eq!(
+            cache.subscriptions(ResourceType::RouteConfiguration),
+            vec!["example-route"],
+        );
+        assert_eq!(
+            cache.versions(ResourceType::RouteConfiguration),
+            collect_kv_str![("example-route", 123)],
+        );
     }
 
     #[test]
     fn test_insert_load_assignment() {
+        let kube_backend = BackendId {
+            service: Service::kube("default", "whatever").unwrap(),
+            port: 7890,
+        };
         let mut cache = Cache::default();
 
-        assert_subscribe_insert(
-            &mut cache,
-            ResourceVec::from_clusters(
-                "123".into(),
-                [
-                    xds_test::cluster!("cluster1.example:8913"),
-                    xds_test::cluster!("cluster2.example:8913"),
-                ],
-            ),
-        );
-
+        // try to insert before being referenced
         assert_insert(cache.insert(ResourceVec::from_load_assignments(
             "123".into(),
             vec![xds_test::cla!(
-                "cluster1.example:8913" => {
+                "whatever.default.svc.cluster.local:7890" => {
+                    "zone1" => ["1.1.1.1"]
+                }
+            )],
+        )));
+        let (resources, dns) = cache.collect();
+        assert!(resources.values().all(|c| c.is_empty()));
+        assert!(dns.is_noop());
+        assert!(cache.data.load_assignments.is_empty());
+
+        // insert a cluster
+        assert_insert(cache.insert(ResourceVec::from_clusters(
+            "123".into(),
+            vec![xds_test::cluster!(kube_backend.name().leak())],
+        )));
+        let _ = cache.collect();
+
+        // try again
+        assert_insert(cache.insert(ResourceVec::from_load_assignments(
+            "123".into(),
+            vec![xds_test::cla!(
+                "whatever.default.svc.cluster.local:7890" => {
                     "zone1" => ["1.1.1.1"]
                 }
             )],
         )));
 
-        assert!(cache.data.listeners.is_empty());
-        assert!(cache.data.route_configs.is_empty());
+        // should be inserted, but won't cause changes
+        let (resources, dns) = cache.collect();
+        assert!(resources.values().all(|c| c.is_empty()));
+        assert!(dns.is_noop());
         assert_eq!(
-            cache.data.clusters.names().collect::<Vec<_>>(),
-            vec!["cluster1.example:8913", "cluster2.example:8913"],
-        );
-        assert_eq!(
-            cache.data.load_assignments.names().collect::<Vec<_>>(),
-            vec!["cluster1.example:8913"],
-        );
-
-        assert_insert(cache.insert(ResourceVec::from_load_assignments(
-            "123".into(),
-            vec![xds_test::cla!(
-                "cluster2.example:8913" => {
-                    "zone2" => ["2.2.2.2"]
-                }
-            )],
-        )));
-
-        assert_eq!(
-            cache.data.clusters.names().collect::<Vec<_>>(),
-            vec!["cluster1.example:8913", "cluster2.example:8913"],
-        );
-        assert_eq!(
-            cache.data.load_assignments.names().collect::<Vec<_>>(),
-            vec!["cluster1.example:8913", "cluster2.example:8913"],
+            cache.versions(ResourceType::ClusterLoadAssignment),
+            collect_kv_str![("whatever.default.svc.cluster.local:7890", "123")]
         );
     }
 
     #[test]
-    fn test_insert_load_assignment_missing_ref() {
+    fn test_remove_listener_explicit() {
         let mut cache = Cache::default();
+        cache.set_wildcard(ResourceType::Cluster, false);
+        cache.set_wildcard(ResourceType::Listener, false);
 
-        assert_subscribe_insert(
-            &mut cache,
-            ResourceVec::from_clusters(
-                "123".into(),
-                [
-                    xds_test::cluster!("cluster1.example:8913"),
-                    xds_test::cluster!("cluster2.example:8913"),
-                ],
-            ),
-        );
-
-        assert_eq!(
-            cache.data.clusters.names().collect::<Vec<_>>(),
-            vec!["cluster1.example:8913", "cluster2.example:8913"],
-        );
-
-        // add a CLA referencing a cluster that doesn't exist. it should just fall on the floor
-        let (changed, errors) = cache.insert(ResourceVec::from_load_assignments(
-            "123".into(),
-            vec![xds_test::cla!(
-                "cluster3.example.svc.cluster.local" => {
-                    "zone2" => ["2.2.2.2"]
-                }
-            )],
-        ));
-        assert!(changed.is_empty());
-        assert!(errors.is_empty());
-    }
-
-    #[test]
-    fn test_insert_deletes_listeners() {
-        let mut cache = Cache::default();
-
-        assert_subscribe_insert(
-            &mut cache,
-            ResourceVec::from_listeners(
-                "123".into(),
-                vec![xds_test::listener!(
-                    "nginx.default.local", "example-route" => [xds_test::vhost!(
-                        "default",
-                        ["nginx.default.local"],
-                        [xds_test::route!(default "nginx.default.local:80")],
-                    )],
-                )],
-            ),
-        );
-
-        assert!(cache.data.listeners.get("nginx.default.local").is_some());
-        assert_eq!(cache.refs.node_count(), 2);
-
-        assert_insert(cache.insert(ResourceVec::from_listeners("123".into(), Vec::new())));
-
-        assert!(cache.data.listeners.is_empty());
-        assert_eq!(
-            cache.refs.node_count(),
-            1,
-            "should still be subscribed to the removed Listener",
-        );
-    }
-
-    #[test]
-    fn test_insert_deletes_clusters() {
-        let mut cache = Cache::default();
-
-        assert_subscribe_insert(
-            &mut cache,
-            ResourceVec::from_clusters(
-                "123".into(),
-                vec![
-                    xds_test::cluster!("cluster1.example:8913"),
-                    xds_test::cluster!("cluster2.example:8913"),
-                ],
-            ),
-        );
-
-        assert_eq!(
-            cache.data.clusters.names().collect::<Vec<_>>(),
-            vec!["cluster1.example:8913", "cluster2.example:8913"],
-        );
-
-        assert_insert(cache.insert(ResourceVec::from_clusters(
-            "123".into(),
-            vec![xds_test::cluster!("cluster2.example:8913")],
-        )));
-
-        assert_eq!(
-            cache.data.clusters.names().collect::<Vec<_>>(),
-            vec!["cluster2.example:8913"],
-        );
-
-        assert_insert(cache.insert(ResourceVec::from_clusters("123".into(), vec![])));
-        assert!(cache.data.clusters.is_empty());
-    }
-
-    #[test]
-    fn test_deletes_keep_subscriptions() {
-        let mut cache = Cache::default();
-
+        // subscribe to two listeners
         cache.subscribe(ResourceType::Listener, "listener.example.svc.cluster.local");
-        cache.subscribe(ResourceType::Cluster, "cluster1.example:8913");
-        cache.subscribe(ResourceType::Cluster, "cluster2.example:8913");
+        cache.subscribe(ResourceType::Listener, "listener.local");
+        let _ = cache.collect();
 
+        // insert a listener -> route -> cluster -> dns chain of configuration
         assert_insert(cache.insert(ResourceVec::from_listeners(
             "123".into(),
             vec![xds_test::listener!(
-                "listener.example.svc.cluster.local", "example-route" => [xds_test::vhost!(
-                    "default",
-                    ["listener.example.svc.cluster.local"],
-                    [
-                        xds_test::route!(header "x-staging" => "cluster2.example:8913"),
-                        xds_test::route!(default "cluster1.example:8913"),
-                    ],
-                )],
+                "listener.example.svc.cluster.local",
+                "example-route",
             )],
         )));
-
-        assert_insert(cache.insert(ResourceVec::from_clusters(
-            "123".into(),
-            vec![
-                xds_test::cluster!("cluster1.example:8913"),
-                xds_test::cluster!("cluster2.example:8913"),
-            ],
-        )));
-
-        assert_eq!(
-            cache.data.clusters.names().collect::<Vec<_>>(),
-            vec!["cluster1.example:8913", "cluster2.example:8913"],
-        );
-
-        // delete everything
-        assert_insert(cache.insert(ResourceVec::from_listeners("123".into(), vec![])));
-        assert_insert(cache.insert(ResourceVec::from_clusters("123".into(), vec![])));
-
-        // subscriptions should still exist, but data should be gone
-        assert!(cache.data.listeners.is_empty());
-        assert!(cache.data.clusters.is_empty());
-        assert_eq!(
-            cache.subscriptions(ResourceType::Listener),
-            vec!["listener.example.svc.cluster.local"],
-        );
-        assert_eq!(
-            cache.subscriptions(ResourceType::Cluster),
-            vec!["cluster1.example:8913", "cluster2.example:8913"],
-        );
-    }
-
-    #[test]
-    fn test_insert_out_of_order() {
-        let mut cache = Cache::default();
-
-        assert_insert(cache.insert(ResourceVec::from_clusters(
-            "123".into(),
-            vec![xds_test::cluster!(
-                inline "cluster1.example.svc.cluster.local" => {
-                "zone1" => ["1.1.1.1", "2.2.2.2"],
-                "zone2" => ["3.3.3.3"]
-            })],
-        )));
-
-        assert!(cache.data.listeners.is_empty());
-        assert!(cache.data.clusters.is_empty());
-
-        assert_subscribe_insert(
-            &mut cache,
-            ResourceVec::from_listeners(
-                "123".into(),
-                vec![xds_test::listener!(
-                    "listener.example.svc.cluster.local", "example-route" => [xds_test::vhost!(
-                        "default",
-                        ["listener.example.svc.cluster.local"],
-                        [xds_test::route!(default "cluster1.example:8913")],
-                    )],
-                )],
-            ),
-        );
-
-        assert!(cache
-            .data
-            .listeners
-            .get("listener.example.svc.cluster.local")
-            .is_some());
-        assert!(cache.data.clusters.is_empty());
-    }
-
-    #[test]
-    fn test_cache_cluster_finds_lb_config_listener() {
-        let mut cache = Cache::default();
-
-        let svc = Service::kube("default", "something")
-            .unwrap()
-            .as_backend_id(8910);
-        let cluster_name = svc.name().leak();
-        let lb_config_name = svc.lb_config_route_name().leak();
-
-        assert_subscribe_insert(
-            &mut cache,
-            ResourceVec::from_listeners(
-                "123".into(),
-                vec![xds_test::listener!(
-                    lb_config_name, lb_config_name => [xds_test::vhost!(
-                        "default-vhost-whatever",
-                        ["*.example.com"],
-                        [xds_test::route!(default ring_hash = "x-user", cluster_name)],
-                    )],
-                )],
-            ),
-        );
-        assert_insert(cache.insert(ResourceVec::from_clusters(
-            "123".into(),
-            vec![xds_test::cluster!(ring_hash cluster_name)],
-        )));
-
-        assert!(
-            {
-                let cluster = cache
-                    .data
-                    .clusters
-                    .get(cluster_name)
-                    .expect("Cache should contain cluster");
-
-                let cluster_data = cluster.data().expect("cluster should have data");
-                matches!(
-                    &cluster_data.backend_lb.config.lb,
-                    LbPolicy::RingHash(params) if !params.hash_params.is_empty(),
-                )
-            },
-            "should have non-empty hash params"
-        );
-    }
-
-    #[test]
-    fn test_cache_cluster_finds_lb_config_route() {
-        let mut cache = Cache::default();
-
-        let svc = Service::kube("default", "something")
-            .unwrap()
-            .as_backend_id(8910);
-        let cluster_name = svc.name().leak();
-        let lb_config_name = svc.lb_config_route_name().leak();
-
-        assert_subscribe_insert(
-            &mut cache,
-            ResourceVec::from_listeners(
-                "123".into(),
-                vec![xds_test::listener!(lb_config_name, lb_config_name)],
-            ),
-        );
-        assert_insert(cache.insert(ResourceVec::from_route_configs(
-            "123".into(),
-            vec![xds_test::route_config!(
-                lb_config_name,
-                [xds_test::vhost!(
-                    "example-vhost",
-                    ["listener.example.svc.cluster.local"],
-                    [xds_test::route!(default ring_hash = "x-user", cluster_name)],
-                )]
-            )],
-        )));
-        assert_insert(cache.insert(ResourceVec::from_clusters(
-            "123".into(),
-            vec![xds_test::cluster!(ring_hash cluster_name)],
-        )));
-
-        assert!(
-            {
-                let cluster = cache
-                    .data
-                    .clusters
-                    .get(cluster_name)
-                    .expect("Cache should contain cluster");
-
-                let cluster_data = cluster.data().expect("cluster should have data");
-                matches!(
-                    &cluster_data.backend_lb.config.lb,
-                    LbPolicy::RingHash(params) if !params.hash_params.is_empty(),
-                )
-            },
-            "should have non-empty hash params"
-        );
-    }
-
-    #[test]
-    fn test_cache_listener_rebuilds_cluster() {
-        let mut cache = Cache::default();
-
-        let svc = Service::kube("default", "something")
-            .unwrap()
-            .as_backend_id(8910);
-        let cluster_name = svc.name().leak();
-        let lb_config_name = svc.lb_config_route_name().leak();
-
-        assert_subscribe_insert(
-            &mut cache,
-            ResourceVec::from_listeners(
-                "123".into(),
-                vec![xds_test::listener!(
-                    "listener.example.svc.cluster.local", "example-route" => [xds_test::vhost!(
-                        "default",
-                        ["listener.example.svc.cluster.local"],
-                        [xds_test::route!(default cluster_name)],
-                    )],
-                )],
-            ),
-        );
-        assert_insert(cache.insert(ResourceVec::from_clusters(
-            "123".into(),
-            vec![xds_test::cluster!(ring_hash cluster_name)],
-        )));
-
-        assert!(
-            {
-                let cluster = cache
-                    .data
-                    .clusters
-                    .get(cluster_name)
-                    .expect("Cache should contain cluster");
-                let cluster_data = cluster.data().expect("cluster should have data");
-
-                matches!(
-                    &cluster_data.backend_lb.config.lb,
-                    LbPolicy::RingHash(params) if params.hash_params.is_empty(),
-                )
-            },
-            "should have empty hash params before Listener insert"
-        );
-
-        assert_insert(cache.insert(ResourceVec::from_listeners(
-            "123".into(),
-            vec![
-                xds_test::listener!(
-                    "listener.example.svc.cluster.local", "example-route" => [xds_test::vhost!(
-                        "default",
-                        ["listener.example.svc.cluster.local"],
-                        [xds_test::route!(default cluster_name)],
-                    )],
-                ),
-                xds_test::listener!(
-                    lb_config_name, lb_config_name => [xds_test::vhost!(
-                        "default",
-                        ["listener.example.svc.cluster.local"],
-                        [xds_test::route!(default ring_hash = "x-user", cluster_name)],
-                    )],
-                ),
-            ],
-        )));
-
-        assert!(
-            {
-                let cluster = cache
-                    .data
-                    .clusters
-                    .get(cluster_name)
-                    .expect("Cache should contain cluster");
-
-                let cluster_data = cluster.data().expect("cluster should have data");
-                matches!(
-                    &cluster_data.backend_lb.config.lb,
-                    LbPolicy::RingHash(params) if !params.hash_params.is_empty(),
-                )
-            },
-            "should have non-empty hash params after Listener insert"
-        );
-    }
-
-    #[test]
-    fn test_cache_route_rebuilds_cluster() {
-        let mut cache = Cache::default();
-
-        let svc = Service::kube("default", "something")
-            .unwrap()
-            .as_backend_id(8910);
-        let cluster_name = svc.name().leak();
-        let lb_config_name = svc.lb_config_route_name().leak();
-
-        assert_subscribe_insert(
-            &mut cache,
-            ResourceVec::from_listeners(
-                "123".into(),
-                vec![xds_test::listener!(lb_config_name, lb_config_name)],
-            ),
-        );
-        assert_insert(cache.insert(ResourceVec::from_route_configs(
-            "123".into(),
-            vec![xds_test::route_config!(
-                lb_config_name,
-                [xds_test::vhost!(
-                    "example-vhost",
-                    ["listener.example.svc.cluster.local"],
-                    [xds_test::route!(default cluster_name),],
-                )]
-            )],
-        )));
-        assert_insert(cache.insert(ResourceVec::from_clusters(
-            "123".into(),
-            vec![xds_test::cluster!(ring_hash cluster_name)],
-        )));
-
-        assert!(
-            {
-                let cluster = cache
-                    .data
-                    .clusters
-                    .get(cluster_name)
-                    .expect("Cache should contain cluster");
-
-                let cluster_data = cluster.data().expect("cluster should have data");
-                matches!(
-                    &cluster_data.backend_lb.config.lb,
-                    LbPolicy::RingHash(params) if params.hash_params.is_empty(),
-                )
-            },
-            "should have empty hash params before the default route has a hash policy"
-        );
-
-        assert_insert(cache.insert(ResourceVec::from_route_configs(
-            "123".into(),
-            vec![xds_test::route_config!(
-                lb_config_name,
-                [xds_test::vhost!(
-                    "example-vhost",
-                    ["listener.example.svc.cluster.local"],
-                    [xds_test::route!(default ring_hash = "x-user", cluster_name)],
-                )]
-            )],
-        )));
-
-        assert!(
-            {
-                let cluster = cache
-                    .data
-                    .clusters
-                    .get(cluster_name)
-                    .expect("Cache should contain cluster");
-
-                let cluster_data = cluster.data().expect("cluster should have data");
-                matches!(
-                    &cluster_data.backend_lb.config.lb,
-                    LbPolicy::RingHash(params) if !params.hash_params.is_empty(),
-                )
-            },
-            "should have non-empty hash params when the default route is updated with a hash policy",
-        );
-    }
-
-    #[test]
-    fn test_cache_gc_simple() {
-        let mut cache = Cache::default();
-
-        assert_subscribe_insert(
-            &mut cache,
-            ResourceVec::from_listeners(
-                "123".into(),
-                vec![xds_test::listener!(
-                    "listener.example.svc.cluster.local:8888",
-                    "example-route"
-                )],
-            ),
-        );
-
         assert_insert(cache.insert(ResourceVec::from_route_configs(
             "123".into(),
             vec![xds_test::route_config!(
                 "example-route",
-                [xds_test::vhost!(
-                    "vhost1.example.svc.cluster.local",
+                vec![xds_test::vhost!(
+                    "a-vhost",
                     ["listener.example.svc.cluster.local"],
-                    [
-                        xds_test::route!(header "x-staging" => "cluster2.example:8913"),
-                        xds_test::route!(default "cluster1.example:8913"),
-                    ],
+                    [xds_test::route!(default "cluster.example:8008")]
                 )]
             )],
         )));
-
         assert_insert(cache.insert(ResourceVec::from_clusters(
             "123".into(),
-            vec![
-                xds_test::cluster!("cluster1.example:8913"),
-                xds_test::cluster!("cluster2.example:8913"),
-            ],
+            vec![xds_test::cluster!("cluster.example:8008")],
         )));
-
-        assert_eq!(
-            cache.data.listeners.names().collect::<Vec<_>>(),
-            vec!["listener.example.svc.cluster.local:8888"],
-        );
-        assert_eq!(
-            cache.data.route_configs.names().collect::<Vec<_>>(),
-            vec!["example-route"],
-        );
-        assert_eq!(
-            cache.data.clusters.names().collect::<Vec<_>>(),
-            vec!["cluster1.example:8913", "cluster2.example:8913"],
-        );
-        assert!(cache.data.load_assignments.is_empty());
-
-        // should have gc refs for everything
-        //
-        // listener, rc,  2 * cluster + 2 * default listener, 2 * cla
-        assert_eq!(cache.refs.node_count(), 8);
-
-        // delete the listener
-        assert_insert(cache.insert(ResourceVec::from_listeners("123".into(), vec![])));
-
-        assert!(cache.data.listeners.is_empty());
-        assert!(cache.data.route_configs.is_empty());
-        assert!(cache.data.clusters.is_empty());
-        assert!(cache.data.load_assignments.is_empty());
-
-        // should have a single ref left for the Listener we subscribed to
-        assert_eq!(cache.refs.node_count(), 1);
-    }
-
-    #[test]
-    fn test_cache_gc_update_rds() {
-        let mut cache = Cache::default();
-
-        // swap the routeconfig for a listener, point to the same clusters
-        assert_subscribe_insert(
-            &mut cache,
-            ResourceVec::from_listeners(
-                "123".into(),
-                vec![xds_test::listener!(
-                    "listener.example.svc.cluster.local:80",
-                    "example-route-1",
+        assert_insert(cache.insert(ResourceVec::from_listeners(
+            "123".into(),
+            vec![xds_test::listener!(
+                "cluster.example.lb.jct:8008",
+                "lb-route" => [xds_test::vhost!(
+                    "lb-vhost",
+                    ["cluster.example.lb.jct:8080"],
+                    [xds_test::route!(default "cluster.example:8008")],
                 )],
-            ),
-        );
-
-        assert_insert(cache.insert(ResourceVec::from_route_configs(
-            "123".into(),
-            vec![xds_test::route_config!(
-                "example-route-1",
-                [xds_test::vhost!(
-                    "whatever",
-                    ["listener.example.svc.cluster.local:80"],
-                    [
-                        xds_test::route!(header "x-staging" => "cluster2.example:8913"),
-                        xds_test::route!(default "cluster1.example:8913"),
-                    ],
-                )]
             )],
         )));
 
-        assert_insert(cache.insert(ResourceVec::from_clusters(
-            "123".into(),
-            vec![
-                xds_test::cluster!("cluster1.example:8913"),
-                xds_test::cluster!("cluster2.example:8913"),
+        // check that the first set of resources makes sense
+        let _ = cache.collect();
+        assert_eq!(
+            cache.versions(ResourceType::Cluster),
+            collect_kv_str![("cluster.example:8008", "123")],
+        );
+        assert_eq!(
+            cache.versions(ResourceType::Listener),
+            collect_kv_str![
+                ("listener.example.svc.cluster.local", "123"),
+                ("cluster.example.lb.jct:8008", "123"),
             ],
-        )));
-
-        // we should have listener -> rc1 -> {cluster1, cluster2}
-        assert_eq!(
-            cache.data.listeners.names().collect::<Vec<_>>(),
-            vec!["listener.example.svc.cluster.local:80"],
         );
         assert_eq!(
-            cache.data.route_configs.names().collect::<Vec<_>>(),
-            vec!["example-route-1"],
+            cache.versions(ResourceType::RouteConfiguration),
+            collect_kv_str![("example-route", "123")],
         );
-        assert_eq!(
-            cache.data.clusters.names().collect::<Vec<_>>(),
-            vec!["cluster1.example:8913", "cluster2.example:8913"],
-        );
-        assert!(cache.data.load_assignments.is_empty());
+        assert!(cache
+            .versions(ResourceType::ClusterLoadAssignment)
+            .is_empty());
 
-        // update the route actions for example-route-1
-        //
-        // should now have listener -> rc1 -> cluster1
-        assert_insert(cache.insert(ResourceVec::from_route_configs(
-            "124".into(),
-            vec![xds_test::route_config!(
-                "example-route-1",
-                [xds_test::vhost!(
-                    "whatever",
-                    ["listener.example.svc.cluster.local:80"],
-                    [xds_test::route!(default "cluster1.example:8913")],
-                )]
-            )],
-        )));
-
-        assert_eq!(
-            cache.data.listeners.names().collect::<Vec<_>>(),
-            vec!["listener.example.svc.cluster.local:80"],
-        );
-        assert_eq!(
-            cache.data.route_configs.names().collect::<Vec<_>>(),
-            vec!["example-route-1"],
-        );
-        assert_eq!(
-            cache.data.clusters.names().collect::<Vec<_>>(),
-            vec!["cluster1.example:8913"],
-        );
-        assert!(cache.data.load_assignments.is_empty());
-    }
-
-    #[test]
-    fn test_cache_gc_pinned() {
-        let mut cache = Cache::default();
-
-        // pinning should let us insert a cluster, but only that cluster
-        cache.subscribe(ResourceType::Cluster, "cluster1.example:8888");
-        assert_insert(cache.insert(ResourceVec::from_clusters(
-            "123".into(),
-            vec![
-                xds_test::cluster!("cluster1.example:8888"),
-                xds_test::cluster!("cluster2.example:8888"),
-            ],
-        )));
-
-        assert!(cache.data.listeners.is_empty());
-        assert_eq!(
-            cache.data.clusters.names().collect::<Vec<_>>(),
-            vec!["cluster1.example:8888"],
-        );
-
-        // add a listener that references both cluster1 and cluster2
-        cache.subscribe(
+        // add the second listener/route-config pointing to the same cluster
+        // and remove the first one
+        cache.remove(
             ResourceType::Listener,
-            "listener.example.svc.cluster.local:8910",
+            &["listener.example.svc.cluster.local".to_string()],
         );
         assert_insert(cache.insert(ResourceVec::from_listeners(
             "123".into(),
             vec![xds_test::listener!(
-                "listener.example.svc.cluster.local:8910", "example-route" => [xds_test::vhost!(
-                    "default",
-                    ["*.listener.example.svc.cluster.local"],
-                    [
-                        xds_test::route!(header "x-staging" => "cluster2.example:8888"),
-                        xds_test::route!(default "cluster1.example:8888"),
-                    ],
+                "listener.local",
+                "better-example-route",
+            )],
+        )));
+        assert_insert(cache.insert(ResourceVec::from_route_configs(
+            "123".into(),
+            vec![xds_test::route_config!(
+                "better-example-route",
+                vec![xds_test::vhost!(
+                    "a-vhost",
+                    ["listener.local"],
+                    [xds_test::route!(default "cluster.example:8008")]
+                )]
+            )],
+        )));
+
+        // should add a remove for the RouteConfig. the cache is still subscribed
+        // to the removed Listener, so it shouldn't appear here.
+        let (resources, dns) = cache.collect();
+        assert_eq!(
+            resources,
+            enum_map::enum_map! {
+                ResourceType::RouteConfiguration => Changes {
+                    removed: collect_str!("example-route"),
+                    ..Default::default()
+                },
+                _ => Changes::default()
+            }
+        );
+        assert!(dns.is_noop());
+
+        // cache should only contain the data versions for what's there, but
+        // should have subscriptions to the original two listeners and the
+        // cluster lb config listener
+        assert_eq!(
+            cache.subscriptions(ResourceType::Listener),
+            vec![
+                "listener.example.svc.cluster.local",
+                "listener.local",
+                "cluster.example.lb.jct:8008",
+            ],
+        );
+
+        assert_eq!(
+            cache.versions(ResourceType::Listener),
+            collect_kv_str![
+                ("listener.local", "123"),
+                ("cluster.example.lb.jct:8008", "123"),
+            ],
+        );
+        assert_eq!(
+            cache.versions(ResourceType::RouteConfiguration),
+            collect_kv_str![("better-example-route", "123")],
+        );
+        assert_eq!(
+            cache.versions(ResourceType::Cluster),
+            collect_kv_str![("cluster.example:8008", "123")],
+        );
+
+        // removing the other listener should drop all data from cache,
+        // but keep both subscriptions.
+        cache.remove(ResourceType::Listener, &["listener.local".to_string()]);
+        let (resources, dns) = cache.collect();
+        assert_eq!(
+            dns,
+            DnsUpdates {
+                add: BTreeSet::new(),
+                remove: [(Hostname::from_static("cluster.example"), 8008)]
+                    .into_iter()
+                    .collect(),
+                sync: false,
+            }
+        );
+
+        assert_eq!(
+            resources,
+            enum_map::enum_map! {
+                ResourceType::Listener => Changes {
+                    removed: collect_str!("cluster.example.lb.jct:8008"),
+                    ..Default::default()
+                },
+                ResourceType::RouteConfiguration => Changes {
+                    removed: collect_str!("better-example-route"),
+                    ..Default::default()
+                },
+                ResourceType::Cluster => Changes {
+                    removed: collect_str!("cluster.example:8008"),
+                    ..Default::default()
+                },
+                ResourceType::ClusterLoadAssignment => Changes::default(),
+            }
+        );
+
+        assert_eq!(
+            cache.subscriptions(ResourceType::Listener),
+            vec!["listener.example.svc.cluster.local", "listener.local"],
+        );
+        for &rtype in ResourceType::all() {
+            assert_eq!(cache.versions(rtype), HashMap::new());
+        }
+    }
+
+    #[test]
+    fn test_remove_listener_wildcard() {
+        let mut cache = Cache::default();
+        cache.set_wildcard(ResourceType::Cluster, true);
+        cache.set_wildcard(ResourceType::Listener, true);
+
+        // subscribe to one listener
+        cache.subscribe(ResourceType::Listener, "listener.example.svc.cluster.local");
+        let _ = cache.collect();
+
+        // insert two listener->route pairs pointing at the same cluster
+        assert_insert(cache.insert(ResourceVec::from_listeners(
+            "123".into(),
+            vec![
+                xds_test::listener!("listener.example.svc.cluster.local", "example-route"),
+                xds_test::listener!("listener.local", "better-example-route"),
+            ],
+        )));
+        assert_insert(cache.insert(ResourceVec::from_route_configs(
+            "123".into(),
+            vec![
+                xds_test::route_config!(
+                    "example-route",
+                    vec![xds_test::vhost!(
+                        "a-vhost",
+                        ["listener.example.svc.cluster.local"],
+                        [xds_test::route!(default "cluster.example:8008")]
+                    )]
+                ),
+                xds_test::route_config!(
+                    "better-example-route",
+                    vec![xds_test::vhost!(
+                        "a-vhost",
+                        ["listener.local"],
+                        [xds_test::route!(default "cluster.example:8008")]
+                    )]
+                ),
+            ],
+        )));
+        assert_insert(cache.insert(ResourceVec::from_clusters(
+            "123".into(),
+            vec![xds_test::cluster!("cluster.example:8008")],
+        )));
+        assert_insert(cache.insert(ResourceVec::from_listeners(
+            "123".into(),
+            vec![xds_test::listener!(
+                "cluster.example.lb.jct:8008",
+                "lb-route" => [xds_test::vhost!(
+                    "lb-vhost",
+                    ["cluster.example.lb.jct:8080"],
+                    [xds_test::route!(default "cluster.example:8008")],
                 )],
             )],
         )));
 
+        // check that the first set of resources makes sense
+        let _ = cache.collect();
         assert_eq!(
-            cache.data.listeners.names().collect::<Vec<_>>(),
-            vec!["listener.example.svc.cluster.local:8910"],
+            cache.versions(ResourceType::Cluster),
+            collect_kv_str![("cluster.example:8008", "123")],
         );
         assert_eq!(
-            cache.data.clusters.names().collect::<Vec<_>>(),
-            vec!["cluster1.example:8888"],
-        );
-
-        // add both clusters
-        assert_insert(cache.insert(ResourceVec::from_clusters(
-            "123".into(),
-            vec![
-                xds_test::cluster!("cluster1.example:8888"),
-                xds_test::cluster!("cluster2.example:8888"),
+            cache.versions(ResourceType::Listener),
+            collect_kv_str![
+                ("listener.local", "123"),
+                ("listener.example.svc.cluster.local", "123"),
+                ("cluster.example.lb.jct:8008", "123"),
             ],
-        )));
-
-        assert_eq!(
-            cache.data.listeners.names().collect::<Vec<_>>(),
-            vec!["listener.example.svc.cluster.local:8910"],
         );
         assert_eq!(
-            cache.data.clusters.names().collect::<Vec<_>>(),
-            vec!["cluster1.example:8888", "cluster2.example:8888"],
+            cache.versions(ResourceType::RouteConfiguration),
+            collect_kv_str![("example-route", "123"), ("better-example-route", "123")],
+        );
+        assert!(cache
+            .versions(ResourceType::ClusterLoadAssignment)
+            .is_empty());
+
+        // remove the explicitly subscribed listener
+        cache.remove(
+            ResourceType::Listener,
+            &["listener.example.svc.cluster.local".to_string()],
         );
 
-        // remove the listener, cluster1 should stay pinned
-        assert_insert(cache.insert(ResourceVec::from_listeners("123".into(), vec![])));
-
-        assert!(cache.data.listeners.is_empty());
+        let (resources, dns) = cache.collect();
         assert_eq!(
-            cache.data.clusters.names().collect::<Vec<_>>(),
-            vec!["cluster1.example:8888"],
+            resources,
+            enum_map::enum_map! {
+                ResourceType::RouteConfiguration => Changes {
+                    removed: collect_str!("example-route"),
+                    ..Default::default()
+                },
+                _ => Changes::default()
+            }
         );
+        assert!(dns.is_noop());
+
+        assert_eq!(
+            cache.subscriptions(ResourceType::Listener),
+            vec![
+                "listener.example.svc.cluster.local",
+                "cluster.example.lb.jct:8008"
+            ],
+        );
+
+        assert_eq!(
+            cache.versions(ResourceType::Listener),
+            collect_kv_str![
+                ("listener.local", "123"),
+                ("cluster.example.lb.jct:8008", "123"),
+            ],
+        );
+        assert_eq!(
+            cache.versions(ResourceType::RouteConfiguration),
+            collect_kv_str![("better-example-route", "123")],
+        );
+        assert_eq!(
+            cache.versions(ResourceType::Cluster),
+            collect_kv_str![("cluster.example:8008", "123")],
+        );
+
+        // removing the wildcard listener should drop the rest of the data
+        cache.remove(ResourceType::Listener, &["listener.local".to_string()]);
+        let (resources, dns) = cache.collect();
+        assert_eq!(
+            dns,
+            DnsUpdates {
+                add: BTreeSet::new(),
+                remove: [(Hostname::from_static("cluster.example"), 8008)]
+                    .into_iter()
+                    .collect(),
+                sync: false,
+            }
+        );
+        assert_eq!(
+            resources,
+            enum_map::enum_map! {
+                ResourceType::Listener => Changes {
+                    removed: collect_str!("cluster.example.lb.jct:8008"),
+                    ..Default::default()
+                },
+                ResourceType::RouteConfiguration => Changes {
+                    removed: collect_str!("better-example-route"),
+                    ..Default::default()
+                },
+                ResourceType::Cluster => Changes {
+                    removed: collect_str!("cluster.example:8008"),
+                    ..Default::default()
+                },
+                ResourceType::ClusterLoadAssignment => Changes::default(),
+            }
+        );
+
+        assert_eq!(
+            cache.subscriptions(ResourceType::Listener),
+            vec!["listener.example.svc.cluster.local"],
+        );
+        assert_eq!(cache.versions(ResourceType::Listener), collect_kv_str![]);
     }
 }
